Fit ARIMA models via formula

arima_tidy is a wrapper around arima that allows you to specify the response variable and any external regressors for an ARIMA model using a data.frame and formula, instead of having to manually generate xreg. This makes the workflow for fitting ARIMA models much closer to the workflow for fitting lm or glm models.

Usage

arima_tidy(
  f,
  data,
  order = c(0L, 0L, 0L),
  seasonal = list(order = c(0L, 0L, 0L), period = NA),
  include.mean = TRUE,
  transform.pars = TRUE,
  fixed = NULL,
  init = NULL,
  method = c("CSS-ML", "ML", "CSS"),
  n.cond,
  SSinit = c("Gardner1980", "Rossignol2011"),
  optim.method = "BFGS",
  optim.control = list(),
  kappa = 1e+06
)

Arguments

f

A formula describing the response variable and any external regressors to be used in the arima model. All variables included in f must be contained in data. The response variable is passed to arima as x and all external regressors are passed as xreg.

data

A data frame or data frame extension (e.g., a tibble) containing all of the variables in the model

order

A specification of the non-seasonal part of the ARIMA model: the three integer components \((p, d, q)\) are the AR order, the degree of differencing, and the MA order.

seasonal

A specification of the seasonal part of the ARIMA model, plus the period. This must be specified by a list with components order and period.

include.mean

Should the ARMA model include a mean/intercept term? The default is TRUE for undifferenced series, and it is ignored for ARIMA models with differencing.

transform.pars

logical; if true, the AR parameters are transformed to ensure that they remain in the region of stationarity. Not used for method = "CSS". For method = "ML", it has been advantageous to set transform.pars = FALSE in some cases, see also fixed.

fixed

optional numeric vector of the same length as the total number of coefficients to be estimated. It should be of the form $$(\phi_1, \ldots, \phi_p, \theta_1, \ldots, \theta_q, \Phi_1, \ldots, \Phi_P, \Theta_1, \ldots, \Theta_Q, \mu), $$ where \(\phi_i\) are the AR coefficients, \(\theta_i\) are the MA coefficients, \(\Phi_i\) are the seasonal AR coefficients, \(\Theta_i\) are the seasonal MA coefficients and \(\mu\) is the intercept term. Note that the \(\mu\) entry is required if and only if include.mean is TRUE. In particular it should not be present if the model is an ARIMA model with differencing.

The entries of the fixed vector should consist of the values at which the user wishes to “fix” the corresponding coefficient, or NA if that coefficient should not be fixed, but estimated.

The argument transform.pars will be set to FALSE if any AR parameters are fixed. A warning will be given if transform.pars is set to (or left at its default) TRUE. It may be wise to set transform.pars = FALSE even when fixing MA parameters, especially at values that cause the model to be nearly non-invertible.

init

optional numeric vector of initial parameter values. Missing values will be filled in, by zeroes except for regression coefficients. Values already specified in fixed will be ignored.

method

fitting method: maximum likelihood or minimize conditional sum-of-squares. The default (unless there are missing values) is to use conditional-sum-of-squares to find starting values, then maximum likelihood. Can be abbreviated.

n.cond

only used if fitting by conditional-sum-of-squares: the number of initial observations to ignore. It will be ignored if less than the maximum lag of an AR term.

SSinit

a string specifying the algorithm to compute the state-space initialization of the likelihood; see KalmanLike for details. Can be abbreviated.

optim.method

The value passed as the method argument to optim.

optim.control

List of control parameters for optim.

kappa

the prior variance (as a multiple of the innovations variance) for the past observations in a differenced model. Do not reduce this.

Value

A model fit object of class "arima_tidy". This is an extension of the Arima model fit produced by arima, so functions such as predict will work identically. The output of arima_tidy contains some additional components that make working with them easier:

formula

The formula used to fit the model

data

The data used to fit the model

xreg

The generated values of xreg used as external regressors

fitted

The fitted values produced by the model. The output of arima contains the residuals, but not the fitted values.

See also

Examples

data("LakeHuron")
LakeHuron_df <- data.frame(lake_level = as.numeric(LakeHuron),
                           date = seq(start(LakeHuron)[[1]],
                                      end(LakeHuron)[[1]],
                                      by = 1))
LakeHuron_df$election_year = LakeHuron_df$date %% 4 == 0

# Basically anything that works in a formula can be used in arima_tidy
arima_tidy(lake_level ~ I(date - 1920), data = LakeHuron_df, order = c(2, 0, 0))
#> 
#> Call:
#> stats::arima(x = resp, order = order, seasonal = seasonal, xreg = xreg, include.mean = include.mean, 
#>     transform.pars = transform.pars, fixed = fixed, init = init, method = method, 
#>     n.cond = n.cond, SSinit = SSinit, optim.method = optim.method, optim.control = optim.control, 
#>     kappa = kappa)
#> 
#> Coefficients:
#>          ar1      ar2  intercept  I(date - 1920)
#>       1.0048  -0.2913   579.0993         -0.0216
#> s.e.  0.0976   0.1004     0.2370          0.0081
#> 
#> sigma^2 estimated as 0.4566:  log likelihood = -101.2,  aic = 212.4

arima_tidy(sqrt(lake_level) ~ date + election_year, data = LakeHuron_df, order = c(2, 0, 0))
#> 
#> Call:
#> stats::arima(x = resp, order = order, seasonal = seasonal, xreg = xreg, include.mean = include.mean, 
#>     transform.pars = transform.pars, fixed = fixed, init = init, method = method, 
#>     n.cond = n.cond, SSinit = SSinit, optim.method = optim.method, optim.control = optim.control, 
#>     kappa = kappa)
#> 
#> Coefficients:
#> Warning: NaNs produced
#>          ar1      ar2  intercept    date  election_yearTRUE
#>       1.0058  -0.2928    24.9299  -5e-04             0.0011
#> s.e.  0.0854      NaN     0.3214   1e-04             0.0020
#> 
#> sigma^2 estimated as 0.0001965:  log likelihood = 278.59,  aic = -545.18

# We can reference variables outside of the `data` argument

shift <- 2000
arima_tidy(sqrt(lake_level) ~ log(date + shift),  data = LakeHuron_df, order = c(2, 0, 0))
#> 
#> Call:
#> stats::arima(x = resp, order = order, seasonal = seasonal, xreg = xreg, include.mean = include.mean, 
#>     transform.pars = transform.pars, fixed = fixed, init = init, method = method, 
#>     n.cond = n.cond, SSinit = SSinit, optim.method = optim.method, optim.control = optim.control, 
#>     kappa = kappa)
#> 
#> Coefficients:
#>          ar1      ar2  intercept  log(date + shift)
#>       1.0046  -0.2917    38.6558            -1.7636
#> s.e.  0.0976   0.1003     5.4524             0.6589
#> 
#> sigma^2 estimated as 0.0001971:  log likelihood = 278.44,  aic = -546.89

# We can handle data with NA's and seasonal data:
data("presidents_df")

arima_tidy(approval_rating ~ 1, data = presidents_df, order = c(1, 0, 0))
#> 
#> Call:
#> stats::arima(x = resp, order = order, seasonal = seasonal, xreg = NULL, include.mean = include.mean, 
#>     transform.pars = transform.pars, fixed = fixed, init = init, method = method, 
#>     n.cond = n.cond, SSinit = SSinit, optim.method = optim.method, optim.control = optim.control, 
#>     kappa = kappa)
#> 
#> Coefficients:
#>          ar1  intercept
#>       0.8242    56.1505
#> s.e.  0.0555     4.6434
#> 
#> sigma^2 estimated as 85.47:  log likelihood = -416.89,  aic = 839.78

# We lose the frequency information when the data is
# stored in a data.frame, so we have to specify
# the period manually for seasonal approaches:
arima_tidy(approval_rating ~ 1, data = presidents_df, order=c(2,0,1),
           seasonal= list(order = c(1,0,0), period = 4),
           fixed=c(NA, NA, 0.5, -0.1, 50), transform.pars=FALSE)
#> 
#> Call:
#> stats::arima(x = resp, order = order, seasonal = seasonal, xreg = NULL, include.mean = include.mean, 
#>     transform.pars = transform.pars, fixed = fixed, init = init, method = method, 
#>     n.cond = n.cond, SSinit = SSinit, optim.method = optim.method, optim.control = optim.control, 
#>     kappa = kappa)
#> 
#> Coefficients:
#>          ar1     ar2  ma1  sar1  intercept
#>       0.2047  0.6269  0.5  -0.1         50
#> s.e.  0.0734  0.0747  0.0   0.0          0
#> 
#> sigma^2 estimated as 84.01:  log likelihood = -416.07,  aic = 838.14