Augment data with information from an arima_tidy object

A method for the augment generic for arima_tidy objects. Takes a model object and a dataset, and adds information about fitted values in the .fitted column and model residuals in the .resid column.

Usage

# S3 method for arima_tidy
augment(x, newdata, ...)

Arguments

x

The result of an arima_tidy fit.

newdata

The data used to originally fit the model as well as any additional data to be used to generate predictions. The rows are assumed to be ordered from earliest to latest. If the originally model was fit on n data points, the first n rows of newdata are assumed to be the data used to fit the model, and any additional rows are used to generate predictions using predict.arima_tidy.

...

Additional arguments. Not used.

Value

newdata with additional columns .fitted and .resid containing the predicted values and the residuals.

Examples

data(presidents_df)
# Fit on the first 110 rows, predict the last 10
fit <- arima_tidy(approval_rating ~ 1,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
# Augment takes the whole dataset (both training and prediction data)
augment(fit, presidents_df)
#>     approval_rating    date election_year  .fitted       .resid
#> 1                NA 1945.25             0       NA           NA
#> 2                87 1945.50             0 69.51369  17.48630754
#> 3                82 1945.75             0 80.96226   1.03773849
#> 4                75 1946.00             0 77.02747  -2.02747477
#> 5                63 1946.25             0 71.51877  -8.51877333
#> 6                50 1946.50             0 62.07529 -12.07528515
#> 7                43 1946.75             0 51.84484  -8.84483963
#> 8                32 1947.00             0 46.33614 -14.33613819
#> 9                35 1947.25             0 37.67961  -2.67960736
#> 10               60 1947.50             0 40.04048  19.95952060
#> 11               54 1947.75             0 59.71441  -5.71441311
#> 12               55 1948.00             1 54.99267   0.00733098
#> 13               36 1948.25             1 55.77963 -19.77962637
#> 14               39 1948.50             1 40.82744  -1.82743675
#> 15               NA 1948.75             1       NA           NA
#> 16               NA 1949.00             0       NA           NA
#> 17               69 1949.25             0 54.92311  14.07689119
#> 18               57 1949.50             0 66.79703  -9.79702924
#> 19               57 1949.75             0 57.35354  -0.35354106
#> 20               51 1950.00             0 57.35354  -6.35354106
#> 21               45 1950.25             0 52.63180  -7.63179698
#> 22               37 1950.50             0 47.91005 -10.91005289
#> 23               46 1950.75             0 41.61439   4.38560590
#> 24               39 1951.00             0 48.69701  -9.69701023
#> 25               36 1951.25             0 43.18831  -7.18830880
#> 26               24 1951.50             0 40.82744 -16.82743675
#> 27               32 1951.75             0 31.38395   0.61605143
#> 28               23 1952.00             1 37.67961 -14.67960736
#> 29               25 1952.25             1 30.59699  -5.59699122
#> 30               32 1952.50             1 32.17091  -0.17090592
#> 31               NA 1952.75             1       NA           NA
#> 32               32 1953.00             0 39.97570  -7.97569599
#> 33               59 1953.25             0 37.67961  21.32039264
#> 34               74 1953.50             0 58.92746  15.07254424
#> 35               75 1953.75             0 70.73182   4.26818402
#> 36               60 1954.00             0 71.51877 -11.51877333
#> 37               71 1954.25             0 59.71441  11.28558689
#> 38               61 1954.50             0 68.37094  -7.37094394
#> 39               71 1954.75             0 60.50137  10.49862954
#> 40               57 1955.00             0 68.37094 -11.37094394
#> 41               71 1955.25             0 57.35354  13.64645894
#> 42               68 1955.50             0 68.37094  -0.37094394
#> 43               79 1955.75             0 66.01007  12.98992810
#> 44               73 1956.00             1 74.66660  -1.66660273
#> 45               76 1956.25             1 69.94486   6.05514136
#> 46               71 1956.50             1 72.30573  -1.30573068
#> 47               67 1956.75             1 68.37094  -1.37094394
#> 48               75 1957.00             0 65.22311   9.77688545
#> 49               79 1957.25             0 71.51877   7.48122667
#> 50               62 1957.50             0 74.66660 -12.66660273
#> 51               63 1957.75             0 61.28833   1.71167219
#> 52               57 1958.00             0 62.07529  -5.07528515
#> 53               60 1958.25             0 57.35354   2.64645894
#> 54               49 1958.50             0 59.71441 -10.71441311
#> 55               48 1958.75             0 51.05788  -3.05788228
#> 56               52 1959.00             0 50.27092   1.72907507
#> 57               57 1959.25             0 53.41875   3.58124568
#> 58               62 1959.50             0 57.35354   4.64645894
#> 59               61 1959.75             0 61.28833  -0.28832781
#> 60               66 1960.00             1 60.50137   5.49862954
#> 61               71 1960.25             1 64.43616   6.56384280
#> 62               62 1960.50             1 68.37094  -6.37094394
#> 63               61 1960.75             1 61.28833  -0.28832781
#> 64               57 1961.00             0 60.50137  -3.50137046
#> 65               72 1961.25             0 57.35354  14.64645894
#> 66               83 1961.50             0 69.15790  13.84209871
#> 67               71 1961.75             0 77.81443  -6.81443212
#> 68               78 1962.00             0 68.37094   9.62905606
#> 69               79 1962.25             0 73.87965   5.12035462
#> 70               71 1962.50             0 74.66660  -3.66660273
#> 71               62 1962.75             0 68.37094  -6.37094394
#> 72               74 1963.00             0 61.28833  12.71167219
#> 73               76 1963.25             0 70.73182   5.26818402
#> 74               64 1963.50             0 72.30573  -8.30573068
#> 75               62 1963.75             0 62.86224  -0.86224250
#> 76               57 1964.00             1 61.28833  -4.28832781
#> 77               80 1964.25             1 57.35354  22.64645894
#> 78               73 1964.50             1 75.45356  -2.45356007
#> 79               69 1964.75             1 69.94486  -0.94485864
#> 80               69 1965.00             0 66.79703   2.20297076
#> 81               71 1965.25             0 66.79703   4.20297076
#> 82               64 1965.50             0 68.37094  -4.37094394
#> 83               69 1965.75             0 62.86224   6.13775750
#> 84               62 1966.00             0 66.79703  -4.79702924
#> 85               63 1966.25             0 61.28833   1.71167219
#> 86               46 1966.50             0 62.07529 -16.07528515
#> 87               56 1966.75             0 48.69701   7.30298977
#> 88               44 1967.00             0 56.56658 -12.56658372
#> 89               44 1967.25             0 47.12310  -3.12309554
#> 90               52 1967.50             0 47.12310   4.87690446
#> 91               38 1967.75             0 53.41875 -15.41875432
#> 92               46 1968.00             1 42.40135   3.59864855
#> 93               36 1968.25             1 48.69701 -12.69701023
#> 94               49 1968.50             1 40.82744   8.17256325
#> 95               35 1968.75             1 51.05788 -16.05788228
#> 96               44 1969.00             0 40.04048   3.95952060
#> 97               59 1969.25             0 47.12310  11.87690446
#> 98               65 1969.50             0 58.92746   6.07254424
#> 99               65 1969.75             0 63.64920   1.35080015
#> 100              56 1970.00             0 63.64920  -7.64919985
#> 101              66 1970.25             0 56.56658   9.43341628
#> 102              53 1970.50             0 64.43616 -11.43615720
#> 103              61 1970.75             0 54.20571   6.79428833
#> 104              52 1971.00             0 60.50137  -8.50137046
#> 105              51 1971.25             0 53.41875  -2.41875432
#> 106              48 1971.50             0 52.63180  -4.63179698
#> 107              54 1971.75             0 50.27092   3.72907507
#> 108              49 1972.00             1 54.99267  -5.99266902
#> 109              49 1972.25             1 51.05788  -2.05788228
#> 110              61 1972.50             1 51.05788   9.94211772
#> 111              NA 1972.75             1 60.50137           NA
#> 112              NA 1973.00             0 60.10897           NA
#> 113              68 1973.25             0 59.80017   8.19983193
#> 114              44 1973.50             0 59.55715 -15.55715390
#> 115              40 1973.75             0 59.36591 -19.36591212
#> 116              27 1974.00             0 59.21541 -32.21541299
#> 117              28 1974.25             0 59.09698 -31.09697660
#> 118              25 1974.50             0 59.00377 -34.00377221
#> 119              24 1974.75             0 58.93042 -34.93042433
#> 120              24 1975.00             0 58.87270 -34.87270267

fit <- arima_tidy(approval_rating ~ election_year,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
augment(fit, presidents_df)
#>     approval_rating    date election_year  .fitted      .resid
#> 1                NA 1945.25             0       NA          NA
#> 2                87 1945.50             0 69.93468  17.0653153
#> 3                82 1945.75             0 81.07760   0.9223990
#> 4                75 1946.00             0 77.15253  -2.1525269
#> 5                63 1946.25             0 71.65742  -8.6574231
#> 6                50 1946.50             0 62.23725 -12.2372452
#> 7                43 1946.75             0 52.03205  -9.0320525
#> 8                32 1947.00             0 46.53695 -14.5369487
#> 9                35 1947.25             0 37.90179  -2.9017857
#> 10               60 1947.50             0 40.25683  19.7431699
#> 11               54 1947.75             0 59.88220  -5.8822008
#> 12               55 1948.00             1 52.05969   2.9403066
#> 13               36 1948.25             1 55.28800 -19.2880028
#> 14               39 1948.50             1 40.37272  -1.3727211
#> 15               NA 1948.75             1       NA          NA
#> 16               NA 1949.00             0       NA          NA
#> 17               69 1949.25             0 56.30450  12.6955032
#> 18               57 1949.50             0 66.94733  -9.9473342
#> 19               57 1949.75             0 57.52716  -0.5271563
#> 20               51 1950.00             0 57.52716  -6.5271563
#> 21               45 1950.25             0 52.81707  -7.8170673
#> 22               37 1950.50             0 48.10698 -11.1069784
#> 23               46 1950.75             0 41.82686   4.1731402
#> 24               39 1951.00             0 48.89199  -9.8919932
#> 25               36 1951.25             0 43.39689  -7.3968894
#> 26               24 1951.50             0 41.04184 -17.0418450
#> 27               32 1951.75             0 31.62167   0.3783329
#> 28               23 1952.00             1 34.78937 -11.7893672
#> 29               25 1952.25             1 30.16753  -5.1675284
#> 30               32 1952.50             1 31.73756   0.2624419
#> 31               NA 1952.75             1       NA          NA
#> 32               32 1953.00             0 41.79519  -9.7951905
#> 33               59 1953.25             0 37.90179  21.0982143
#> 34               74 1953.50             0 59.09719  14.9028141
#> 35               75 1953.75             0 70.87241   4.1275917
#> 36               60 1954.00             0 71.65742 -11.6574231
#> 37               71 1954.25             0 59.88220  11.1177992
#> 38               61 1954.50             0 68.51736  -7.5173638
#> 39               71 1954.75             0 60.66722  10.3327844
#> 40               57 1955.00             0 68.51736 -11.5173638
#> 41               71 1955.25             0 57.52716  13.4728437
#> 42               68 1955.50             0 68.51736  -0.5173638
#> 43               79 1955.75             0 66.16232  12.8376806
#> 44               73 1956.00             1 71.68506   1.3149360
#> 45               76 1956.25             1 69.41827   6.5817303
#> 46               71 1956.50             1 71.77331  -0.7733141
#> 47               67 1956.75             1 67.84824  -0.8482400
#> 48               75 1957.00             0 67.82060   7.1794008
#> 49               79 1957.25             0 71.65742   7.3425769
#> 50               62 1957.50             0 74.79748 -12.7974824
#> 51               63 1957.75             0 61.45223   1.5477696
#> 52               57 1958.00             0 62.23725  -5.2372452
#> 53               60 1958.25             0 57.52716   2.4728437
#> 54               49 1958.50             0 59.88220 -10.8822008
#> 55               48 1958.75             0 51.24704  -3.2470377
#> 56               52 1959.00             0 50.46202   1.5379771
#> 57               57 1959.25             0 53.60208   3.3979178
#> 58               62 1959.50             0 57.52716   4.4728437
#> 59               61 1959.75             0 61.45223  -0.4522304
#> 60               66 1960.00             1 57.55480   8.4452029
#> 61               71 1960.25             1 63.92317   7.0768341
#> 62               62 1960.50             1 67.84824  -5.8482400
#> 63               61 1960.75             1 60.78311   0.2168934
#> 64               57 1961.00             0 63.11051  -6.1105102
#> 65               72 1961.25             0 57.52716  14.4728437
#> 66               83 1961.50             0 69.30238  13.6976213
#> 67               71 1961.75             0 77.93754  -6.9375417
#> 68               78 1962.00             0 68.51736   9.4826362
#> 69               79 1962.25             0 74.01247   4.9875324
#> 70               71 1962.50             0 74.79748  -3.7974824
#> 71               62 1962.75             0 68.51736  -6.5173638
#> 72               74 1963.00             0 61.45223  12.5477696
#> 73               76 1963.25             0 70.87241   5.1275917
#> 74               64 1963.50             0 72.44244  -8.4424380
#> 75               62 1963.75             0 63.02226  -1.0222601
#> 76               57 1964.00             1 58.33981  -1.3398120
#> 77               80 1964.25             1 56.85803  23.1419675
#> 78               73 1964.50             1 74.91337  -1.9133734
#> 79               69 1964.75             1 69.41827  -0.4182697
#> 80               69 1965.00             0 69.39063  -0.3906288
#> 81               71 1965.25             0 66.94733   4.0526658
#> 82               64 1965.50             0 68.51736  -4.5173638
#> 83               69 1965.75             0 63.02226   5.9777399
#> 84               62 1966.00             0 66.94733  -4.9473342
#> 85               63 1966.25             0 61.45223   1.5477696
#> 86               46 1966.50             0 62.23725 -16.2372452
#> 87               56 1966.75             0 48.89199   7.1080068
#> 88               44 1967.00             0 56.74214 -12.7421415
#> 89               44 1967.25             0 47.32196  -3.3219636
#> 90               52 1967.50             0 47.32196   4.6780364
#> 91               38 1967.75             0 53.60208 -15.6020822
#> 92               46 1968.00             1 39.49946   6.5005438
#> 93               36 1968.25             1 48.22287 -12.2228694
#> 94               49 1968.50             1 40.37272   8.6272789
#> 95               35 1968.75             1 50.57791 -15.5779139
#> 96               44 1969.00             0 42.70012   1.2998752
#> 97               59 1969.25             0 47.32196  11.6780364
#> 98               65 1969.50             0 59.09719   5.9028141
#> 99               65 1969.75             0 63.80727   1.1927251
#> 100              56 1970.00             0 63.80727  -7.8072749
#> 101              66 1970.25             0 56.74214   9.2578585
#> 102              53 1970.50             0 64.59229 -11.5922897
#> 103              61 1970.75             0 54.38710   6.6129030
#> 104              52 1971.00             0 60.66722  -8.6672156
#> 105              51 1971.25             0 53.60208  -2.6020822
#> 106              48 1971.50             0 52.81707  -4.8170673
#> 107              54 1971.75             0 50.46202   3.5379771
#> 108              49 1972.00             1 52.05969  -3.0596934
#> 109              49 1972.25             1 50.57791  -1.5779139
#> 110              61 1972.50             1 50.57791  10.4220861
#> 111              NA 1972.75             1 59.99809          NA
#> 112              NA 1973.00             0 62.32400          NA
#> 113              68 1973.25             0 61.70657   6.2934268
#> 114              44 1973.50             0 61.22189 -17.2218860
#> 115              40 1973.75             0 60.84140 -20.8413994
#> 116              27 1974.00             0 60.54271 -33.5427118
#> 117              28 1974.25             0 60.30824 -32.3082376
#> 118              25 1974.50             0 60.12417 -35.1241718
#> 119              24 1974.75             0 59.97968 -35.9796775
#> 120              24 1975.00             0 59.86625 -35.8662473