Combine the most recent data points from a sequence of data frames

In statistical surveillance, "surveillance residuals" are defined as the difference between observed data for a new time period and the predictions from a model fit using previous time periods. The key difference from ordinary model residuals is that surveillance residuals are never calculated for data points used to fit the model. In theory, examining surveillance residuals should allow statistical methods to detect outbreaks or clusters more rapidly, since the model used to predict baseline estimates is not affected by any current outbreaks. In practice, it is still an open question whether using surveillance residuals over ordinary residuals provides any appreciable benefit.

calculate_surveillance_residuals takes a set of data frames and combines the most recent time period or time periods from each one. If the data frames contain the estimates from sequentially fitting a model, then the new set of data frames contains the surveillance residuals from the series of model fits.

Usage

calculate_surveillance_residuals(
  list_of_dataframes,
  window_time_ids,
  split_ids,
  grow_length = FALSE,
  include_init = FALSE,
  check_space_ids = TRUE
)

Arguments

list_of_dataframes

A list of data frames or data frame extensions, organized from oldest to newest. Each data frame must includes columns 'id_space' and 'id_time' that are each a set of consecutive integers. The column 'id_time' identifies the time point associated with each row, with lower numbers being earlier and higher numbers being later. The column 'id_space' identifies the spatial location or area associated with each row, but no meaning or order is ascribed to the values.

window_time_ids

A vector the same length as list_of_dataframes. Each entry should be the value of id_time associated with each data frame in list_of_dataframes. For example, if list_of_dataframes contains model output, window_time_ids would be the time point when each model was run, or the latest timepoint included in each model.

split_ids

A vector the same length of list_of_dataframes. Each entry should be the value of id_time that marks the beginning of the prediction data in each data frame.

grow_length

Boolean. Should the size of the returned dataframes increase as the number of available surveillance predictions increase?

include_init

Boolean. Should the output include time points that were always used in fitting the model and never had predictions generated for them? If include_init is TRUE and grow_length is false, then the dimensions of the returned data frames are all identical to the size of the input data frames.

check_space_ids

Boolean. Should the function throw an error if the values of id_space in each data frame are not the same. Defaults to TRUE, and you should use caution before overriding.

Value

A list of data frames giving the surveillance predictions for every entry in list_of_dataframes. If grow_length = FALSE, then surveillance predictions will never be reported for a time point not included in the corresponding dataframe in list_of_dataframes. If grow_lenth = TRUE, then each data frame will contain all of the surveillance predictions calculated up to that point. If include_init = TRUE, then each data frame will also include data points that never had any predictions generated for them.

Details

Unlike ordinary model predictions, which are reported for all data points, surveillance predictions are only reported for points that were not used in calculating the model fit. For example, if a model was run at time \(t\) and the time points \(0\) to \((t-1)\) were used to calculate the model fit, but model predictions were generated for time points \(0\) through \(t\), then the only surveillance predictions would be at time point \(t\).

While calculate_surveillance_residuals can be used on any list of data frames, it is designed to used on a series of predictions from a model fit sequentially on a growing set of data. Given a list of ordinary model predictions, from the same model fit sequentially on different sets of data, convert_to_surveillance combines surveillance predictions from multiple model fits.

If the same time period has predictions generated for it more than once, only the most recent value is reported.

Examples

library("scanstatistics")
library("magrittr")
library("broom")
data(NM_popcas)

# Generate a series of model predictions, each time including the data from another year:
spacetime_data <- NM_popcas %>%
  dplyr::mutate(id_time = year - min(year),
                id_space = as.numeric(factor(county)))

windowed_data <- spacetime_data %>%
  window_idtime(min_train = 5, max_train = 5, n_predict = 1)

model_res <- windowed_data %>%
  rowmute(training_data = prepare_training_data(curr_data, count, split_id),
          model_fit = glm(count ~ year, family = poisson(),
                          offset = log(population), data = training_data),
          model_predictions = extract_yhat(model_fit, newdata = curr_data))
# The sequential model predictions are now stored in the model_res$model_predictions

# Use convert_to_surveillance to take the last time period
# from each model prediction
calculate_surveillance_residuals(model_res$model_predictions,
                                 model_res$window_time_id,
                                 model_res$split_id)
#> [[1]]
#> # A tibble: 32 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 22 more rows
#> 
#> [[2]]
#> # A tibble: 64 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 54 more rows
#> 
#> [[3]]
#> # A tibble: 96 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 86 more rows
#> 
#> [[4]]
#> # A tibble: 128 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 118 more rows
#> 
#> [[5]]
#> # A tibble: 160 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 150 more rows
#> 
#> [[6]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 182 more rows
#> 
#> [[7]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1980 bernalillo     416461    22       7        1 16.1   
#>  2  1980 catron           2639     0       7        2  0.102 
#>  3  1980 chaves          51928     2       7        3  2.01  
#>  4  1980 colfax          14049     0       7        4  0.544 
#>  5  1980 curry           41846     0       7        5  1.62  
#>  6  1980 debaca           2448     1       7        6  0.0948
#>  7  1980 donaana         96111     4       7        7  3.72  
#>  8  1980 eddy            50760     0       7        8  1.96  
#>  9  1980 grant           27310     2       7        9  1.06  
#> 10  1980 guadalupe        4394     0       7       10  0.170 
#> # ... with 182 more rows
#> 
#> [[8]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1981 bernalillo     425133    21       8        1  19.3  
#>  2  1981 catron           2657     0       8        2   0.121
#>  3  1981 chaves          52810     3       8        3   2.40 
#>  4  1981 colfax          14124     0       8        4   0.641
#>  5  1981 curry           41887     0       8        5   1.90 
#>  6  1981 debaca           2437     0       8        6   0.111
#>  7  1981 donaana         99704     5       8        7   4.53 
#>  8  1981 eddy            51595     2       8        8   2.34 
#>  9  1981 grant           27670     0       8        9   1.26 
#> 10  1981 guadalupe        4338     0       8       10   0.197
#> # ... with 182 more rows
#> 
#> [[9]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1982 bernalillo     433476    21       9        1 15.4   
#>  2  1982 catron           2668     0       9        2  0.0949
#>  3  1982 chaves          53643     3       9        3  1.91  
#>  4  1982 colfax          14155     1       9        4  0.503 
#>  5  1982 curry           41977     6       9        5  1.49  
#>  6  1982 debaca           2425     0       9        6  0.0862
#>  7  1982 donaana        103450     4       9        7  3.68  
#>  8  1982 eddy            52235     5       9        8  1.86  
#>  9  1982 grant           27968     1       9        9  0.994 
#> 10  1982 guadalupe        4288     0       9       10  0.152 
#> # ... with 182 more rows
#> 
#> [[10]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1983 bernalillo     441490    26      10        1 17.9   
#>  2  1983 catron           2673     0      10        2  0.108 
#>  3  1983 chaves          54428     4      10        3  2.21  
#>  4  1983 colfax          14143     0      10        4  0.574 
#>  5  1983 curry           42115     1      10        5  1.71  
#>  6  1983 debaca           2413     0      10        6  0.0979
#>  7  1983 donaana        107350     2      10        7  4.35  
#>  8  1983 eddy            52681     0      10        8  2.14  
#>  9  1983 grant           28202     0      10        9  1.14  
#> 10  1983 guadalupe        4244     0      10       10  0.172 
#> # ... with 182 more rows
#> 
#> [[11]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1984 bernalillo     449175    17      11        1  21.7  
#>  2  1984 catron           2671     0      11        2   0.129
#>  3  1984 chaves          55164     4      11        3   2.67 
#>  4  1984 colfax          14089     0      11        4   0.681
#>  5  1984 curry           42302     0      11        5   2.04 
#>  6  1984 debaca           2400     0      11        6   0.116
#>  7  1984 donaana        111404     5      11        7   5.38 
#>  8  1984 eddy            52932     2      11        8   2.56 
#>  9  1984 grant           28373     1      11        9   1.37 
#> 10  1984 guadalupe        4206     0      11       10   0.203
#> # ... with 182 more rows
#> 
#> [[12]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1985 bernalillo     456531    28      12        1  19.3  
#>  2  1985 catron           2663     1      12        2   0.112
#>  3  1985 chaves          55851     4      12        3   2.36 
#>  4  1985 colfax          13991     2      12        4   0.590
#>  5  1985 curry           42537     1      12        5   1.79 
#>  6  1985 debaca           2387     0      12        6   0.101
#>  7  1985 donaana        115613    10      12        7   4.88 
#>  8  1985 eddy            52989     3      12        8   2.23 
#>  9  1985 grant           28481     2      12        9   1.20 
#> 10  1985 guadalupe        4173     0      12       10   0.176
#> # ... with 182 more rows
#> 
#> [[13]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1986 bernalillo     463558    27      13        1  26.6  
#>  2  1986 catron           2648     1      13        2   0.152
#>  3  1986 chaves          56490     2      13        3   3.24 
#>  4  1986 colfax          13851     1      13        4   0.794
#>  5  1986 curry           42821     2      13        5   2.46 
#>  6  1986 debaca           2374     0      13        6   0.136
#>  7  1986 donaana        119975     9      13        7   6.88 
#>  8  1986 eddy            52851     1      13        8   3.03 
#>  9  1986 grant           28526     0      13        9   1.64 
#> 10  1986 guadalupe        4147     1      13       10   0.238
#> # ... with 182 more rows
#> 

# Repeat with grow_length = FALSE, init = TRUE,
# so the dimensions of the output match the dimensions of the input
calculate_surveillance_residuals(model_res$model_predictions,
                                 model_res$window_time_id,
                                 model_res$split_id,
                                 grow_length = FALSE,
                                 include_init = TRUE)
#> [[1]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1974 bernalillo     357520    16       1        1  15.8  
#>  2  1974 catron           2393     0       1        2   0.106
#>  3  1974 chaves          45616     3       1        3   2.02 
#>  4  1974 colfax          12700     1       1        4   0.561
#>  5  1974 curry           42613     2       1        5   1.88 
#>  6  1974 debaca           2506     0       1        6   0.111
#>  7  1974 donaana         77790     3       1        7   3.44 
#>  8  1974 eddy            41667     1       1        8   1.84 
#>  9  1974 grant           23816     0       1        9   1.05 
#> 10  1974 guadalupe        4856     0       1       10   0.215
#> # ... with 182 more rows
#> 
#> [[2]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1975 bernalillo     368166    16       2        1  15.7  
#>  2  1975 catron           2450     0       2        2   0.105
#>  3  1975 chaves          46789     3       2        3   2.00 
#>  4  1975 colfax          13032     1       2        4   0.556
#>  5  1975 curry           42364     3       2        5   1.81 
#>  6  1975 debaca           2498     0       2        6   0.107
#>  7  1975 donaana         80459     2       2        7   3.43 
#>  8  1975 eddy            43669     1       2        8   1.86 
#>  9  1975 grant           24556     0       2        9   1.05 
#> 10  1975 guadalupe        4764     0       2       10   0.203
#> # ... with 182 more rows
#> 
#> [[3]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1976 bernalillo     378483    16       3        1  15.6  
#>  2  1976 catron           2501     0       3        2   0.103
#>  3  1976 chaves          47914     1       3        3   1.98 
#>  4  1976 colfax          13322     1       3        4   0.550
#>  5  1976 curry           42164     3       3        5   1.74 
#>  6  1976 debaca           2489     0       3        6   0.103
#>  7  1976 donaana         83281     1       3        7   3.44 
#>  8  1976 eddy            45476     1       3        8   1.88 
#>  9  1976 grant           25234     1       3        9   1.04 
#> 10  1976 guadalupe        4678     0       3       10   0.193
#> # ... with 182 more rows
#> 
#> [[4]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1977 bernalillo     388471    15       4        1 15.5   
#>  2  1977 catron           2545     0       4        2  0.101 
#>  3  1977 chaves          48990     1       4        3  1.95  
#>  4  1977 colfax          13568     0       4        4  0.541 
#>  5  1977 curry           42011     4       4        5  1.68  
#>  6  1977 debaca           2479     0       4        6  0.0988
#>  7  1977 donaana         86258     4       4        7  3.44  
#>  8  1977 eddy            47089     3       4        8  1.88  
#>  9  1977 grant           25847     1       4        9  1.03  
#> 10  1977 guadalupe        4598     0       4       10  0.183 
#> # ... with 182 more rows
#> 
#> [[5]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1978 bernalillo     398130    18       5        1 15.3   
#>  2  1978 catron           2583     0       5        2  0.0995
#>  3  1978 chaves          50018     1       5        3  1.93  
#>  4  1978 colfax          13771     0       5        4  0.531 
#>  5  1978 curry           41908     3       5        5  1.61  
#>  6  1978 debaca           2469     0       5        6  0.0951
#>  7  1978 donaana         89388     5       5        7  3.44  
#>  8  1978 eddy            48507     1       5        8  1.87  
#>  9  1978 grant           26398     0       5        9  1.02  
#> 10  1978 guadalupe        4524     0       5       10  0.174 
#> # ... with 182 more rows
#> 
#> [[6]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1979 bernalillo     407460    19       6        1 15.2   
#>  2  1979 catron           2614     0       6        2  0.0974
#>  3  1979 chaves          50997     5       6        3  1.90  
#>  4  1979 colfax          13932     1       6        4  0.519 
#>  5  1979 curry           41852     1       6        5  1.56  
#>  6  1979 debaca           2459     0       6        6  0.0916
#>  7  1979 donaana         92673     4       6        7  3.45  
#>  8  1979 eddy            49731     2       6        8  1.85  
#>  9  1979 grant           26885     1       6        9  1.00  
#> 10  1979 guadalupe        4456     0       6       10  0.166 
#> # ... with 182 more rows
#> 
#> [[7]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1980 bernalillo     416461    22       7        1 16.1   
#>  2  1980 catron           2639     0       7        2  0.102 
#>  3  1980 chaves          51928     2       7        3  2.01  
#>  4  1980 colfax          14049     0       7        4  0.544 
#>  5  1980 curry           41846     0       7        5  1.62  
#>  6  1980 debaca           2448     1       7        6  0.0948
#>  7  1980 donaana         96111     4       7        7  3.72  
#>  8  1980 eddy            50760     0       7        8  1.96  
#>  9  1980 grant           27310     2       7        9  1.06  
#> 10  1980 guadalupe        4394     0       7       10  0.170 
#> # ... with 182 more rows
#> 
#> [[8]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1981 bernalillo     425133    21       8        1  19.3  
#>  2  1981 catron           2657     0       8        2   0.121
#>  3  1981 chaves          52810     3       8        3   2.40 
#>  4  1981 colfax          14124     0       8        4   0.641
#>  5  1981 curry           41887     0       8        5   1.90 
#>  6  1981 debaca           2437     0       8        6   0.111
#>  7  1981 donaana         99704     5       8        7   4.53 
#>  8  1981 eddy            51595     2       8        8   2.34 
#>  9  1981 grant           27670     0       8        9   1.26 
#> 10  1981 guadalupe        4338     0       8       10   0.197
#> # ... with 182 more rows
#> 
#> [[9]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1982 bernalillo     433476    21       9        1 15.4   
#>  2  1982 catron           2668     0       9        2  0.0949
#>  3  1982 chaves          53643     3       9        3  1.91  
#>  4  1982 colfax          14155     1       9        4  0.503 
#>  5  1982 curry           41977     6       9        5  1.49  
#>  6  1982 debaca           2425     0       9        6  0.0862
#>  7  1982 donaana        103450     4       9        7  3.68  
#>  8  1982 eddy            52235     5       9        8  1.86  
#>  9  1982 grant           27968     1       9        9  0.994 
#> 10  1982 guadalupe        4288     0       9       10  0.152 
#> # ... with 182 more rows
#> 
#> [[10]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1983 bernalillo     441490    26      10        1 17.9   
#>  2  1983 catron           2673     0      10        2  0.108 
#>  3  1983 chaves          54428     4      10        3  2.21  
#>  4  1983 colfax          14143     0      10        4  0.574 
#>  5  1983 curry           42115     1      10        5  1.71  
#>  6  1983 debaca           2413     0      10        6  0.0979
#>  7  1983 donaana        107350     2      10        7  4.35  
#>  8  1983 eddy            52681     0      10        8  2.14  
#>  9  1983 grant           28202     0      10        9  1.14  
#> 10  1983 guadalupe        4244     0      10       10  0.172 
#> # ... with 182 more rows
#> 
#> [[11]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1984 bernalillo     449175    17      11        1  21.7  
#>  2  1984 catron           2671     0      11        2   0.129
#>  3  1984 chaves          55164     4      11        3   2.67 
#>  4  1984 colfax          14089     0      11        4   0.681
#>  5  1984 curry           42302     0      11        5   2.04 
#>  6  1984 debaca           2400     0      11        6   0.116
#>  7  1984 donaana        111404     5      11        7   5.38 
#>  8  1984 eddy            52932     2      11        8   2.56 
#>  9  1984 grant           28373     1      11        9   1.37 
#> 10  1984 guadalupe        4206     0      11       10   0.203
#> # ... with 182 more rows
#> 
#> [[12]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1985 bernalillo     456531    28      12        1  19.3  
#>  2  1985 catron           2663     1      12        2   0.112
#>  3  1985 chaves          55851     4      12        3   2.36 
#>  4  1985 colfax          13991     2      12        4   0.590
#>  5  1985 curry           42537     1      12        5   1.79 
#>  6  1985 debaca           2387     0      12        6   0.101
#>  7  1985 donaana        115613    10      12        7   4.88 
#>  8  1985 eddy            52989     3      12        8   2.23 
#>  9  1985 grant           28481     2      12        9   1.20 
#> 10  1985 guadalupe        4173     0      12       10   0.176
#> # ... with 182 more rows
#> 
#> [[13]]
#> # A tibble: 192 x 7
#>     year county     population count id_time id_space .fitted
#>    <int> <fct>           <int> <int>   <int>    <dbl>   <dbl>
#>  1  1986 bernalillo     463558    27      13        1  26.6  
#>  2  1986 catron           2648     1      13        2   0.152
#>  3  1986 chaves          56490     2      13        3   3.24 
#>  4  1986 colfax          13851     1      13        4   0.794
#>  5  1986 curry           42821     2      13        5   2.46 
#>  6  1986 debaca           2374     0      13        6   0.136
#>  7  1986 donaana        119975     9      13        7   6.88 
#>  8  1986 eddy            52851     1      13        8   3.03 
#>  9  1986 grant           28526     0      13        9   1.64 
#> 10  1986 guadalupe        4147     1      13       10   0.238
#> # ... with 182 more rows
#> 

# We can use the function inside mutate to add a column to our data frame
model_res %>%
  dplyr::mutate(surveillance_predictions =
                   calculate_surveillance_residuals(model_predictions,
                                                    window_time_id,
                                                    split_id,
                                                    grow_length = FALSE,
                                                    include_init = TRUE))
#> # A tibble: 13 x 8
#>    window_time_id window_s~1 split~2 curr_d~3 traini~4 model~5 model_~6 survei~7
#>             <int> <list>       <dbl> <list>   <list>   <list>  <list>   <list>  
#>  1              6 <dbl [32]>       6 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  2              7 <dbl [32]>       7 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  3              8 <dbl [32]>       8 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  4              9 <dbl [32]>       9 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  5             10 <dbl [32]>      10 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  6             11 <dbl [32]>      11 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  7             12 <dbl [32]>      12 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  8             13 <dbl [32]>      13 <tibble> <tibble> <glm>   <tibble> <tibble>
#>  9             14 <dbl [32]>      14 <tibble> <tibble> <glm>   <tibble> <tibble>
#> 10             15 <dbl [32]>      15 <tibble> <tibble> <glm>   <tibble> <tibble>
#> 11             16 <dbl [32]>      16 <tibble> <tibble> <glm>   <tibble> <tibble>
#> 12             17 <dbl [32]>      17 <tibble> <tibble> <glm>   <tibble> <tibble>
#> 13             18 <dbl [32]>      18 <tibble> <tibble> <glm>   <tibble> <tibble>
#> # ... with abbreviated variable names 1: window_space_id, 2: split_id,
#> #   3: curr_data, 4: training_data, 5: model_fit, 6: model_predictions,
#> #   7: surveillance_predictions