Merge entries into a list of length 1

Take a vector or list and combine the entries into a list of length 1.

Usage

collapse(vec, combine_dfs = TRUE, unique_only = FALSE)

Arguments

vec

An atomic vector or a list.

combine_dfs

Logical. Whether columns containtaining data frames should be combined into a single longer data frame (combine_dfs = TRUE) or left as a list of data frames (combine_dfs = FALSE)

unique_only

Logical. Whether the entries in a list should contain the values from all rows, including repeats (unique_only = FALSE) or if only unique values should be reported (unique_only = TRUE). Does not apply to data frames if combine_dfs = TRUE.

Value

A list of length one. If combine_dfs = FALSE, unique_only = FALSE, then collapse()

returns a list(vec). If combine_dfs = TRUE, and vec is a list of data frames, then collapse() returns list(bind_rows(vec)). If unique_only = TRUE, then duplicate entries of vec are removed.

Details

collapse() is a helper function used by collapse_all, but can also be used in conjunction with summarize to reduce the number of rows in a data frame without losing information.

Examples

vec <- c(1, 2, 3, 3, 4)
collapse(vec)
#> [[1]]
#> [1] 1 2 3 3 4
#> 

collapse(vec, unique_only = TRUE)
#> [[1]]
#> [1] 1 2 3 4
#> 

df1 <- data.frame(x = c(1, 2, 3), y = c("A", "B", "C"))
df2 <- data.frame(x = c(4, 5), y= c("D", "E"))
collapse(list(df1, df2), combine_dfs = TRUE)
#> [[1]]
#>   x y
#> 1 1 A
#> 2 2 B
#> 3 3 C
#> 4 4 D
#> 5 5 E
#> 

# collapse can be used together with summarize:
library("dplyr")
#> 
#> Attaching package: 'dplyr'
#> The following object is masked from 'package:statsurv':
#> 
#>     collapse
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
library("tidyr")
#> 
#> Attaching package: 'tidyr'
#> The following object is masked from 'package:magrittr':
#> 
#>     extract
mtcars_nested <- mtcars %>%
  group_by(cyl) %>%
  nest()

# Generate two models for mpg with different predicting variables
model_fits <- tidyr::crossing(mtcars_nested, model_var = c("wt", "hp")) %>%
  rowwise() %>%
  mutate(model = list(lm(paste0("mpg ~ ", model_var), data = data)),
         predictions = list(broom::augment(model)))

# Combine the results from each model separately
model_fits %>%
  group_by(model_var) %>%
  summarize(models = statsurv::collapse(model),
            predictions = statsurv::collapse(predictions, combine_dfs = TRUE))
#> # A tibble: 2 x 3
#>   model_var models     predictions      
#>   <chr>     <list>     <list>           
#> 1 hp        <list [3]> <tibble [32 x 8]>
#> 2 wt        <list [3]> <tibble [32 x 8]>