Merge rows from each group into a list

collapse_all() takes a data frame and for each column, combines all the entries for each group into a list. Other parameters control whether data frames should be merged into a single longer data frame, whether only unique values should be reported, or whether columns should be transformed back into atomic columns where possible.

Usage

collapse_all(
  df,
  combine_dfs = TRUE,
  unique_only = FALSE,
  unlist_scalars = TRUE,
  .groups = NULL
)

Arguments

df

A data frame or data frame extension. Can be grouped.

combine_dfs

Logical. Whether columns containtaining data frames should be combined into a single longer data frame (combine_dfs = TRUE) or left as a list of data frames (combine_dfs = FALSE)

unique_only

Logical. Whether the entries in a list should contain the values from all rows, including repeats (unique_only = FALSE) or if only unique values should be reported (unique_only = TRUE). Does not apply to data frames if combine_dfs = TRUE.

unlist_scalars

Logical. After columns are reduced to a single row, should columns containing scalar values be converted from list columns into regular columns.

.groups

Grouping structure of the result. Can be "drop_last", "drop", "keep", or "rowwise". See summarize for details.

Value

An object of the same type as df, with one row per group in the original data frame. Columns other than grouping variables are transformed into list columns. If unlist_scalars = TRUE, some of the list columns may be converted back into regular columns.

Details

While collapse_all can be used on any data frame, it's designed around data frames that having a sequential series of model fits. Each model might be associated with some parameters, such as what date and location it corresponds to, some data, and some predictions. If you have a separate model for each location, it's natural to want to combine the predictions from each location into a larger dataset that you can look at together. collapse_all simplifies the process of combining the data and predictions from each model fit into a single data frame. collapse_all works with grouping variables, so if you have model fits for multiple locations from multiple different models, it's easy to combine the predictions from each model separately.

One way to think about collapse_all is that it's halfway between summarize and nest. Like summarize, it combines all the rows for each group into a single row. Like nest, it stores all the data (unlike unique_only = TRUE), instead of reporting a single summary value for each group. Unlike nest though, collapse_all works on each column individually, and it stores the data as a list instead of as a data frame.

In fact, collapse_all is a wrapper around summarize (not nest), and a call to collapse_all can always be replaced with a call to summarize, with assistance from the helper functions collapse and unlist_scalars:

df 
   collapse_all()

is equivalent to

df 
  summarize(across(everything()), .fns = collapse, combined_dfs = TRUE) 
  unlist_scalars()

Examples

library(tidyr)
library(dplyr)

mtcars_nested <- mtcars %>%
  group_by(cyl) %>%
  nest()

# Generate two models for mpg with different predicting variables
model_fits <- tidyr::crossing(mtcars_nested, model_var = c("wt", "hp")) %>%
  rowwise() %>%
  mutate(model = list(lm(paste0("mpg ~ ", model_var), data = data)),
         predictions = list(broom::augment(model)))

# Combine the results from each model separately
model_fits %>%
  group_by(model_var) %>%
  collapse_all()
#> # A tibble: 2 x 5
#>   model_var cyl       data               model      predictions      
#>   <chr>     <list>    <list>             <list>     <list>           
#> 1 hp        <dbl [3]> <tibble [32 x 10]> <list [3]> <tibble [32 x 8]>
#> 2 wt        <dbl [3]> <tibble [32 x 10]> <list [3]> <tibble [32 x 8]>