Create zones from spatial coordinates

Scan-type alarm functions work by computing an alarm function over a very large number of spatial zones, where each zone consists of one of more neighboring spatial locations. Zones can be many sizes, from a single location up to an aggregation of max_k locations, and each location can belong to multiple zones. create_zones aggregates individual spatial locations into a set of zones using a k-nearest neighbors approach. For each location, there will be a zone containing the location itself, a zone containing the location and its nearest neighbor, a zone containing the location and its two nearest-neighbors, up to the maximum size.

Usage

create_zones(sf, max_k, min_k = 1)

Arguments

sf

A simple-features (sf) collection describing the spatial locations to combine into zones. Each row of sf should represent a unique spatial location.

max_k

The maximum number of locations to be included in any one zone.

min_k

Optional. The minimum number of locations to be included in any one zone.

Value

A list of integer vectors, specifying which locations belong to which zones. The integers identify locations based on which row in sf they are associated with.

Examples

library("sf")
#> Linking to GEOS 3.9.1, GDAL 3.2.1, PROJ 7.2.1; sf_use_s2() is TRUE
data("NM_county_sf")
create_zones(NM_county_sf, max_k = 5, min_k = 2)
#> [[1]]
#> [1]  1 32
#> 
#> [[2]]
#> [1]  1 23 32
#> 
#> [[3]]
#> [1]  1 23 26 32
#> 
#> [[4]]
#> [1]  1 23 26 30 32
#> 
#> [[5]]
#> [1] 2 9
#> 
#> [[6]]
#> [1]  2  9 28
#> 
#> [[7]]
#> [1]  2  9 27 28
#> 
#> [[8]]
#> [1]  2  9 27 28 32
#> 
#> [[9]]
#> [1] 3 8
#> 
#> [[10]]
#> [1]  3  8 14
#> 
#> [[11]]
#> [1]  3  6  8 14
#> 
#> [[12]]
#> [1]  3  6  8 14 22
#> 
#> [[13]]
#> [1]  4 18
#> 
#> [[14]]
#> [1]  4 18 29
#> 
#> [[15]]
#> [1]  4 18 29 31
#> 
#> [[16]]
#> [1]  4 11 18 29 31
#> 
#> [[17]]
#> [1]  5 20
#> 
#> [[18]]
#> [1]  5 20 22
#> 
#> [[19]]
#> [1]  5  6 20 22
#> 
#> [[20]]
#> [1]  5  6 10 20 22
#> 
#> [[21]]
#> [1]  6 10
#> 
#> [[22]]
#> [1]  6 10 22
#> 
#> [[23]]
#> [1]  5  6 10 22
#> 
#> [[24]]
#> [1]  3  5  6 10 22
#> 
#> [[25]]
#> [1]  7 16
#> 
#> [[26]]
#> [1]  7 16 27
#> 
#> [[27]]
#> [1]  7 16 19 27
#> 
#> [[28]]
#> [1]  7  9 16 19 27
#> 
#> [[29]]
#> [1]  8 13
#> 
#> [[30]]
#> [1]  3  8 13
#> 
#> [[31]]
#> [1]  3  8 13 19
#> 
#> [[32]]
#> [1]  3  8 13 14 19
#> 
#> [[33]]
#> [1]  9 16
#> 
#> [[34]]
#> [1]  9 12 16
#> 
#> [[35]]
#> [1]  9 12 16 27
#> 
#> [[36]]
#> [1]  2  9 12 16 27
#> 
#> [[37]]
#> [1]  6 10 25
#> 
#> [[38]]
#> [1]  6 10 25 30
#> 
#> [[39]]
#> [1]  6 10 20 25 30
#> 
#> [[40]]
#> [1] 11 31
#> 
#> [[41]]
#> [1] 11 20 31
#> 
#> [[42]]
#> [1] 11 20 25 31
#> 
#> [[43]]
#> [1] 11 18 20 25 31
#> 
#> [[44]]
#> [1] 12 16
#> 
#> [[45]]
#> [1]  7  9 12 16
#> 
#> [[46]]
#> [1]  7  9 12 16 27
#> 
#> [[47]]
#> [1]  3  8 13 22
#> 
#> [[48]]
#> [1]  3  6  8 13 22
#> 
#> [[49]]
#> [1]  3 14
#> 
#> [[50]]
#> [1]  3 14 30
#> 
#> [[51]]
#> [1]  3  6 14 30
#> 
#> [[52]]
#> [1]  3  6 14 19 30
#> 
#> [[53]]
#> [1] 15 26
#> 
#> [[54]]
#> [1] 15 23 26
#> 
#> [[55]]
#> [1] 15 21 23 26
#> 
#> [[56]]
#> [1]  1 15 21 23 26
#> 
#> [[57]]
#> [1]  7  9 16
#> 
#> [[58]]
#> [1] 17 24
#> 
#> [[59]]
#> [1] 17 23 24
#> 
#> [[60]]
#> [1]  1 17 23 24
#> 
#> [[61]]
#> [1]  1 17 23 24 32
#> 
#> [[62]]
#> [1] 18 25
#> 
#> [[63]]
#> [1]  4 18 25
#> 
#> [[64]]
#> [1]  4 18 25 29
#> 
#> [[65]]
#> [1]  4 11 18 25 29
#> 
#> [[66]]
#> [1]  7 19
#> 
#> [[67]]
#> [1]  7 14 19
#> 
#> [[68]]
#> [1]  7  8 14 19
#> 
#> [[69]]
#> [1]  3  7  8 14 19
#> 
#> [[70]]
#> [1]  5 11 20
#> 
#> [[71]]
#> [1]  5  6 11 20
#> 
#> [[72]]
#> [1]  5  6 10 11 20
#> 
#> [[73]]
#> [1] 15 21
#> 
#> [[74]]
#> [1] 15 21 23
#> 
#> [[75]]
#> [1] 15 21 23 29
#> 
#> [[76]]
#> [1] 15 21 23 26 29
#> 
#> [[77]]
#> [1]  5 22
#> 
#> [[78]]
#> [1]  5  6 22
#> 
#> [[79]]
#> [1]  3  5  6 22
#> 
#> [[80]]
#> [1]  3  5  6 20 22
#> 
#> [[81]]
#> [1] 15 23
#> 
#> [[82]]
#> [1]  1 15 23
#> 
#> [[83]]
#> [1]  1 15 23 26
#> 
#> [[84]]
#> [1] 17 21 24
#> 
#> [[85]]
#> [1] 17 21 23 24
#> 
#> [[86]]
#> [1] 15 17 21 23 24
#> 
#> [[87]]
#> [1] 10 18 25
#> 
#> [[88]]
#> [1] 10 11 18 25
#> 
#> [[89]]
#> [1] 10 11 18 25 26
#> 
#> [[90]]
#> [1]  1 15 26
#> 
#> [[91]]
#> [1]  1 15 23 26 30
#> 
#> [[92]]
#> [1]  7 27
#> 
#> [[93]]
#> [1]  7 27 28
#> 
#> [[94]]
#> [1]  7 16 27 28
#> 
#> [[95]]
#> [1]  7  9 16 27 28
#> 
#> [[96]]
#> [1] 28 32
#> 
#> [[97]]
#> [1] 27 28 32
#> 
#> [[98]]
#> [1]  1 27 28 32
#> 
#> [[99]]
#> [1]  1 27 28 30 32
#> 
#> [[100]]
#> [1]  4 29
#> 
#> [[101]]
#> [1]  4 18 21 29
#> 
#> [[102]]
#> [1]  4 15 18 21 29
#> 
#> [[103]]
#> [1]  1 30
#> 
#> [[104]]
#> [1]  1 30 32
#> 
#> [[105]]
#> [1]  1 26 30 32
#> 
#> [[106]]
#> [1]  1 10 26 30 32
#> 
#> [[107]]
#> [1]  4 11 31
#> 
#> [[108]]
#> [1]  4 11 18 31
#> 
#> [[109]]
#> [1]  4 11 18 20 31
#> 
#> [[110]]
#> [1]  1 28 32
#> 
#> [[111]]
#> [1]  1 28 30 32
#> 
#> [[112]]
#> [1]  1 23 28 30 32
#> 
# 112 zones created from 32 locations