Extract fitted values from a model object

A consistent method for extracting fitted values from a model fit object. Generally a wrapper around augment from the broom package, but one that always returns the values on the scale of the response variable. extract_yhat also covers arima and INLA models that augment does not. extract_yhat is a generic, meaning that the method invoked depends on what type of model is passed as the first argument.

Usage

extract_yhat(fit, newdata, ...)

# S3 method for merMod
extract_yhat(fit, newdata, ...)

# S3 method for inla
extract_yhat(fit, newdata, se_fit = FALSE, ...)

Arguments

fit

A model fit object, such as that generated by lm or glm.

newdata

A data frame containing the data that should be used to predict the fitted values. Must have all of the covariates used by the model fit object to calculate fitted values.

...

Additional arguments passed into other methods

se_fit

If TRUE, also calculate the standard errors of the fitted values. Not available for all methods.

Value

The input data frame newdata with the added column .fitted containing the predicted values. If se_fit = TRUE, then the data frame also contains the column .se.fit with the errors of the predicted values. Some models may also return additional columns, such as .resid. All columns added by extract_yhat or augment start with a dot.

Details

Where possible, extract_yhat passes its arguments to augment, calling augment(x = fit, newdata = newdata, type.predict = "response"). In some cases, such as for glmer models, extract_yhat takes additional steps to better handle NA values in the original dataset used to fit the model. extract_yhat can also extract fitted values from arima models or models from the INLA package that do not have augment methods. See below for details on each individual method.

Method Details

lm, glm: Identical to augment(x = fit, newdata = newdata, type.predict = "response")

lmer, glmer: Starts by calling augment, but removes the columns ".mu", ".offset", ".sqrtXwt", ".sqrtrwt", ".weights", ".wtres", ".gam", and ".eta" from the result. This is because these columns aren't necessary for statistical surveillance, and because they have errors when the original data frame contains NA's. In addition, extract_yhat gives an error if offset is included as an option in glmer. As a workaround, offsets can be specified as a term in the model formula, via offset(VAR)

INLA: Uses the linear predictor calculated by INLA and the link function to calculated fitted values. The model must be fit with the option control.predictor=list(compute = TRUE) in order for extract_yhat to work. Can be called with the option se_fit = TRUE, which gives standard errors for the fitted values, but note that this can be extremely slow.

Arima models: See extract_yhat.Arima

See also

Examples

library("scanstatistics")
library("broom")
library("broom.mixed")
data(NM_popcas)

# Fit a model to all the data, and then extract predictions:
fit <- glm(count ~ year,
           family = poisson(link = "log"),
           offset = log(population),
           data = NM_popcas[1:600, ])

# Then use extract_yhat to get out predictions for our observed variable:
extract_yhat(fit, NM_popcas, se_fit = TRUE)
#> # A tibble: 608 x 6
#>     year county     population count .fitted .se.fit
#>    <int> <fct>           <int> <int>   <dbl>   <dbl>
#>  1  1973 bernalillo     353813    16    13.7   0.851
#>  2  1974 bernalillo     357520    16    14.0   0.807
#>  3  1975 bernalillo     368166    16    14.7   0.778
#>  4  1976 bernalillo     378483    16    15.4   0.745
#>  5  1977 bernalillo     388471    15    16.1   0.710
#>  6  1978 bernalillo     398130    18    16.8   0.674
#>  7  1979 bernalillo     407460    19    17.4   0.641
#>  8  1980 bernalillo     416461    22    18.1   0.613
#>  9  1981 bernalillo     425133    21    18.8   0.593
#> 10  1982 bernalillo     433476    21    19.5   0.587
#> # ... with 598 more rows

# We can also fit glmer models
fit_glmer <- lme4::glmer(count ~ I(year - 1980) + (1 | county) + offset(log(population)),
                         family = "poisson",
                         data = NM_popcas[1:600, ])
extract_yhat(fit_glmer, NM_popcas, se_fit = TRUE)
#> # A tibble: 608 x 5
#>     year county     population count .fitted
#>    <int> <fct>           <int> <int>   <dbl>
#>  1  1973 bernalillo     353813    16    14.9
#>  2  1974 bernalillo     357520    16    15.3
#>  3  1975 bernalillo     368166    16    16.0
#>  4  1976 bernalillo     378483    16    16.8
#>  5  1977 bernalillo     388471    15    17.5
#>  6  1978 bernalillo     398130    18    18.2
#>  7  1979 bernalillo     407460    19    19.0
#>  8  1980 bernalillo     416461    22    19.8
#>  9  1981 bernalillo     425133    21    20.5
#> 10  1982 bernalillo     433476    21    21.3
#> # ... with 598 more rows

if (FALSE) {
  # We can also use INLA models if available:
  # All the data points to generate predictions for much be included in the
  # data arg of the INLA model
  fit_inla <- INLA::inla(count ~ year + f(county, model = "iid"),
                         family = "poisson",
                         control.family = list(link = "log"),
                         offset = log(population),
                         # This option is required to use extract_yhat
                         control.predictor = list(compute = TRUE),
                         data = NM_popcas)

  extract_yhat(fit_inla, NM_popcas)
}