Extract fitted values from an ARIMA model

Methods for extracting fitted values from ARIMA models, including those fit using stats::arima, forecast::Arima, or arima_tidy.

Compared to other types of models, extrating fitted values from Arima models is rather fiddly. In particular, the order of rows in newdata and newxreg is crucially important. If the ARIMA model was originally fit on n points, extract_yhat assumes that the first n rows of newdata and newxreg correspond exactly to the data used to fit the model. The number of additional rows in newdata is taken as the number of steps ahead that predictions should be generated.

Usage

# S3 method for Arima
extract_yhat(fit, newdata, resp_var, newxreg = NULL, ...)

# S3 method for forecast_ARIMA
extract_yhat(fit, newdata, newxreg = NULL, ...)

# S3 method for arima_tidy
extract_yhat(fit, newdata, ...)

Arguments

fit

An ARIMA model fit object.

newdata

A data frame containing the data that should be used to predict the fitted values. Must have all of the covariates used by the model fit object to calculate fitted values. The first n rows should be exactly the data used to fit the model, and any additional rows should correspond to the number of steps ahead that predictions should be generated.

resp_var

The column in newdata containing the time series data the model was fitted on. Only used for stats::arima models. The column should be specified as bare text (no quotes).

newxreg

A matrix of covariate information containing both the values of xreg used to originally fit the model, and any new values of xreg to be used for prediction. IMPORTANT: Unlike predict.Arima, newxreg must contain both the old and new values of xreg.

...

Additional arguments passed into other methods

Value

The input data frame newdata with the added column .fitted containing the predicted values. Some models may also return additional columns, such as .resid. All columns added by extract_yhat or augment start with a dot.

Details

Behind the scenes, extract_yhat is performing two different tasks: extracting already fitted values from the ARIMA model, and forecasting future values of the response variable. For most other models, these two tasks are the same, since once the model is fit, the output of one data point is not dependent on any other data points. The fact that the predictions of an ARIMA model depend on the order of the data points drives a lot of the oddities of extract_yhat for ARIMA models.

Extracting already fitted values is simple for models fit with the forecast::Arima or arima_tidy functions, because the fitted values are stored in the model fit object. Models fit with stats::arima do not store this information, and so this information must be supplied to extract_yhat by through the parameters newdata and resp_var.

extract_yhat can only be used to predict fitted values conditional on the data used to fit the model. For forecast::Arima and stats::arima, any covariates necessary to generate these forecasts must be provided as the last rows in newxreg. For arima_tidy, the covariates must be provided as the last rows in newdata. Although newdata is note used for generating predictions for forecast::Ariam or stats::arima, it is still a required parameter to the functions, so that it can be used to report the final result. This is done so that the output from extract_yhat is consistent with the output from augment and the broader Tidyverse, which uses data frames as the key interface for passing data between functions.

For a version of arima that is based entirely on data frames, see arima_tidy.

See also

Examples

data("presidents_df")
fit_stats_arima <- stats::arima(presidents_df$approval_rating, order = c(2, 0, 0))
extract_yhat(fit_stats_arima, presidents_df, resp_var = approval_rating)
#>     approval_rating    date election_year  .fitted      .resid
#> 1                NA 1945.25             0       NA          NA
#> 2                87 1945.50             0 69.88907  17.1109291
#> 3                82 1945.75             0 81.73735   0.2626502
#> 4                75 1946.00             0 78.84617  -3.8461656
#> 5                63 1946.25             0 73.14549 -10.1454892
#> 6                50 1946.50             0 63.58330 -13.5832995
#> 7                43 1946.75             0 52.63291  -9.6329134
#> 8                32 1947.00             0 45.86112 -13.8611174
#> 9                35 1947.25             0 37.01767  -2.0176744
#> 10               60 1947.50             0 37.70113  22.2988750
#> 11               54 1947.75             0 56.07146  -2.0714616
#> 12               55 1948.00             1 55.10623  -0.1062303
#> 13               36 1948.25             1 55.02164 -19.0216372
#> 14               39 1948.50             1 41.49934  -2.4993405
#> 15               NA 1948.75             1       NA          NA
#> 16               NA 1949.00             0       NA          NA
#> 17               69 1949.25             0 51.54767  17.4523293
#> 18               57 1949.50             0 65.54538  -8.5453846
#> 19               57 1949.75             0 58.46748  -1.4674798
#> 20               51 1950.00             0 56.86080  -5.8608004
#> 21               45 1950.25             0 52.54832  -7.5483204
#> 22               37 1950.50             0 47.43250 -10.4325007
#> 23               46 1950.75             0 40.87919   5.1208123
#> 24               39 1951.00             0 46.27679  -7.2767881
#> 25               36 1951.25             0 42.45057  -6.4505710
#> 26               24 1951.50             0 39.35710 -15.3571013
#> 27               32 1951.75             0 30.33047   1.6695286
#> 28               23 1952.00             1 34.47377 -11.4737654
#> 29               25 1952.25             1 29.07616  -4.0761650
#> 30               32 1952.50             1 29.30865   2.6913512
#> 31               NA 1952.75             1       NA          NA
#> 32               32 1953.00             0 36.40042  -4.4004227
#> 33               59 1953.25             0 35.68754  23.3124628
#> 34               74 1953.50             0 54.95105  19.0489549
#> 35               75 1953.75             0 69.34727   5.6527263
#> 36               60 1954.00             0 72.07437 -12.0743696
#> 37               71 1954.25             0 61.42706   9.5729405
#> 38               61 1954.50             0 67.32492  -6.3249236
#> 39               71 1954.75             0 61.61025   9.3897536
#> 40               57 1955.00             0 67.45881 -10.4588136
#> 41               71 1955.25             0 58.73526  12.2647403
#> 42               68 1955.50             0 66.92325   1.0767462
#> 43               79 1955.75             0 66.64147  12.3585269
#> 44               73 1956.00             1 74.14602  -1.1460166
#> 45               76 1956.25             1 71.30633   4.6936740
#> 46               71 1956.50             1 72.65923  -1.6592263
#> 47               67 1956.75             1 69.46716  -2.4671628
#> 48               75 1957.00             0 65.92273   9.0772736
#> 49               79 1957.25             0 71.13714   7.8628601
#> 50               62 1957.50             0 75.08325 -13.0832462
#> 51               63 1957.75             0 63.40011  -0.4001126
#> 52               57 1958.00             0 61.84273  -4.8427302
#> 53               60 1958.25             0 57.66414   2.3358599
#> 54               49 1958.50             0 59.01704 -10.0170404
#> 55               48 1958.75             0 51.51250  -3.5124969
#> 56               52 1959.00             0 49.32096   2.6790392
#> 57               57 1959.25             0 52.06206   4.9379425
#> 58               62 1959.50             0 56.19135   5.8086493
#> 59               61 1959.75             0 60.45453   0.5454662
#> 60               66 1960.00             1 60.40524   5.5947632
#> 61               71 1960.25             1 63.86508   7.1349198
#> 62               62 1960.50             1 68.12826  -6.1282633
#> 63               61 1960.75             1 62.32899  -1.3289930
#> 64               57 1961.00             0 60.40524  -3.4052368
#> 65               72 1961.25             0 57.39636  14.6036398
#> 66               83 1961.50             0 67.64200  15.3579995
#> 67               71 1961.75             0 77.55656  -6.5565630
#> 68               78 1962.00             0 70.40439   7.5956076
#> 69               79 1962.25             0 73.82894   5.1710603
#> 70               71 1962.50             0 75.48492  -4.4849160
#> 71               62 1962.75             0 69.86883  -7.8688326
#> 72               74 1963.00             0 62.32899  11.6710070
#> 73               76 1963.25             0 69.74894   6.2510565
#> 74               64 1963.50             0 72.79312  -8.7931163
#> 75               62 1963.75             0 64.43594  -2.4359361
#> 76               57 1964.00             1 61.39176  -4.3917634
#> 77               80 1964.25             1 57.53025  22.4697498
#> 78               73 1964.50             1 73.39197  -0.3919738
#> 79               69 1964.75             1 71.44022  -2.4402159
#> 80               69 1965.00             0 67.62800   1.3720004
#> 81               71 1965.25             0 67.09244   3.9075602
#> 82               64 1965.50             0 68.52993  -4.5299332
#> 83               69 1965.75             0 63.76649   5.2335136
#> 84               62 1966.00             0 66.42299  -4.4229901
#> 85               63 1966.25             0 62.06121   0.9387869
#> 86               46 1966.50             0 61.84273 -15.8427302
#> 87               56 1966.75             0 49.75793   6.2420732
#> 88               44 1967.00             0 54.66926 -10.6692643
#> 89               44 1967.25             0 47.38320  -3.3832038
#> 90               52 1967.50             0 45.77652   6.2234756
#> 91               38 1967.75             0 51.52650 -13.5264978
#> 92               46 1968.00             1 42.53516   3.4648360
#> 93               36 1968.25             1 46.41068 -10.4106781
#> 94               49 1968.50             1 40.29433   8.7056690
#> 95               35 1968.75             1 48.29914 -13.2991382
#> 96               44 1969.00             0 39.97725   4.0227459
#> 97               59 1969.25             0 44.57151  14.4284851
#> 98               65 1969.50             0 56.55772   8.4422756
#> 99               65 1969.75             0 62.87855   2.1214463
#> 100              56 1970.00             0 63.68189  -7.6818934
#> 101              66 1970.25             0 57.21317   8.7868267
#> 102              53 1970.50             0 63.19563 -10.1956305
#> 103              61 1970.75             0 55.19082   5.8091767
#> 104              52 1971.00             0 59.20023  -7.2002273
#> 105              51 1971.25             0 53.80263  -2.8026269
#> 106              48 1971.50             0 51.87887  -3.8788707
#> 107              54 1971.75             0 49.58874   4.4112593
#> 108              49 1972.00             1 53.49955  -4.4995509
#> 109              49 1972.25             1 50.70916  -1.7091572
#> 110              61 1972.50             1 50.03971  10.9602925
#> 111              NA 1972.75             1       NA          NA
#> 112              NA 1973.00             0       NA          NA
#> 113              68 1973.25             0 60.98400   7.0159974
#> 114              44 1973.50             0 65.62758 -21.6275794
#> 115              40 1973.75             0 48.98988  -8.9898832
#> 116              27 1974.00             0 42.90154 -15.9015377
#> 117              28 1974.25             0 33.02227  -5.0222712
#> 118              25 1974.50             0 32.00045  -7.0004486
#> 119              24 1974.75             0 29.97810  -5.9780985
#> 120              24 1975.00             0 28.85768  -4.8576820

fit_forecast  <- forecast::Arima(presidents_df$approval_rating, order = c(2, 0, 0))
extract_yhat(fit_forecast, presidents_df)
#>     approval_rating    date election_year  .fitted
#> 1                NA 1945.25             0       NA
#> 2                87 1945.50             0 69.88907
#> 3                82 1945.75             0 81.73735
#> 4                75 1946.00             0 78.84617
#> 5                63 1946.25             0 73.14549
#> 6                50 1946.50             0 63.58330
#> 7                43 1946.75             0 52.63291
#> 8                32 1947.00             0 45.86112
#> 9                35 1947.25             0 37.01767
#> 10               60 1947.50             0 37.70113
#> 11               54 1947.75             0 56.07146
#> 12               55 1948.00             1 55.10623
#> 13               36 1948.25             1 55.02164
#> 14               39 1948.50             1 41.49934
#> 15               NA 1948.75             1       NA
#> 16               NA 1949.00             0       NA
#> 17               69 1949.25             0 51.54767
#> 18               57 1949.50             0 65.54538
#> 19               57 1949.75             0 58.46748
#> 20               51 1950.00             0 56.86080
#> 21               45 1950.25             0 52.54832
#> 22               37 1950.50             0 47.43250
#> 23               46 1950.75             0 40.87919
#> 24               39 1951.00             0 46.27679
#> 25               36 1951.25             0 42.45057
#> 26               24 1951.50             0 39.35710
#> 27               32 1951.75             0 30.33047
#> 28               23 1952.00             1 34.47377
#> 29               25 1952.25             1 29.07616
#> 30               32 1952.50             1 29.30865
#> 31               NA 1952.75             1       NA
#> 32               32 1953.00             0 36.40042
#> 33               59 1953.25             0 35.68754
#> 34               74 1953.50             0 54.95105
#> 35               75 1953.75             0 69.34727
#> 36               60 1954.00             0 72.07437
#> 37               71 1954.25             0 61.42706
#> 38               61 1954.50             0 67.32492
#> 39               71 1954.75             0 61.61025
#> 40               57 1955.00             0 67.45881
#> 41               71 1955.25             0 58.73526
#> 42               68 1955.50             0 66.92325
#> 43               79 1955.75             0 66.64147
#> 44               73 1956.00             1 74.14602
#> 45               76 1956.25             1 71.30633
#> 46               71 1956.50             1 72.65923
#> 47               67 1956.75             1 69.46716
#> 48               75 1957.00             0 65.92273
#> 49               79 1957.25             0 71.13714
#> 50               62 1957.50             0 75.08325
#> 51               63 1957.75             0 63.40011
#> 52               57 1958.00             0 61.84273
#> 53               60 1958.25             0 57.66414
#> 54               49 1958.50             0 59.01704
#> 55               48 1958.75             0 51.51250
#> 56               52 1959.00             0 49.32096
#> 57               57 1959.25             0 52.06206
#> 58               62 1959.50             0 56.19135
#> 59               61 1959.75             0 60.45453
#> 60               66 1960.00             1 60.40524
#> 61               71 1960.25             1 63.86508
#> 62               62 1960.50             1 68.12826
#> 63               61 1960.75             1 62.32899
#> 64               57 1961.00             0 60.40524
#> 65               72 1961.25             0 57.39636
#> 66               83 1961.50             0 67.64200
#> 67               71 1961.75             0 77.55656
#> 68               78 1962.00             0 70.40439
#> 69               79 1962.25             0 73.82894
#> 70               71 1962.50             0 75.48492
#> 71               62 1962.75             0 69.86883
#> 72               74 1963.00             0 62.32899
#> 73               76 1963.25             0 69.74894
#> 74               64 1963.50             0 72.79312
#> 75               62 1963.75             0 64.43594
#> 76               57 1964.00             1 61.39176
#> 77               80 1964.25             1 57.53025
#> 78               73 1964.50             1 73.39197
#> 79               69 1964.75             1 71.44022
#> 80               69 1965.00             0 67.62800
#> 81               71 1965.25             0 67.09244
#> 82               64 1965.50             0 68.52993
#> 83               69 1965.75             0 63.76649
#> 84               62 1966.00             0 66.42299
#> 85               63 1966.25             0 62.06121
#> 86               46 1966.50             0 61.84273
#> 87               56 1966.75             0 49.75793
#> 88               44 1967.00             0 54.66926
#> 89               44 1967.25             0 47.38320
#> 90               52 1967.50             0 45.77652
#> 91               38 1967.75             0 51.52650
#> 92               46 1968.00             1 42.53516
#> 93               36 1968.25             1 46.41068
#> 94               49 1968.50             1 40.29433
#> 95               35 1968.75             1 48.29914
#> 96               44 1969.00             0 39.97725
#> 97               59 1969.25             0 44.57151
#> 98               65 1969.50             0 56.55772
#> 99               65 1969.75             0 62.87855
#> 100              56 1970.00             0 63.68189
#> 101              66 1970.25             0 57.21317
#> 102              53 1970.50             0 63.19563
#> 103              61 1970.75             0 55.19082
#> 104              52 1971.00             0 59.20023
#> 105              51 1971.25             0 53.80263
#> 106              48 1971.50             0 51.87887
#> 107              54 1971.75             0 49.58874
#> 108              49 1972.00             1 53.49955
#> 109              49 1972.25             1 50.70916
#> 110              61 1972.50             1 50.03971
#> 111              NA 1972.75             1       NA
#> 112              NA 1973.00             0       NA
#> 113              68 1973.25             0 60.98400
#> 114              44 1973.50             0 65.62758
#> 115              40 1973.75             0 48.98988
#> 116              27 1974.00             0 42.90154
#> 117              28 1974.25             0 33.02227
#> 118              25 1974.50             0 32.00045
#> 119              24 1974.75             0 29.97810
#> 120              24 1975.00             0 28.85768
# We can extract fitted values and make predictions at the same time
fit <- stats::arima(presidents_df$approval_rating[1:110], order = c(2, 0, 0))
extract_yhat(fit, presidents_df, resp_var = approval_rating)
#>     approval_rating    date election_year  .fitted       .resid
#> 1                NA 1945.25             0       NA           NA
#> 2                87 1945.50             0 69.89757  17.10242948
#> 3                82 1945.75             0 81.05837   0.94162800
#> 4                75 1946.00             0 78.51376  -3.51376042
#> 5                63 1946.25             0 73.07179 -10.07178667
#> 6                50 1946.50             0 63.96582 -13.96581983
#> 7                43 1946.75             0 53.47389 -10.47388902
#> 8                32 1947.00             0 46.89598 -14.89597549
#> 9                35 1947.25             0 38.46601  -3.46601029
#> 10               60 1947.50             0 38.93210  21.06790201
#> 11               54 1947.75             0 56.25812  -2.25811610
#> 12               55 1948.00             1 55.75192  -0.75191811
#> 13               36 1948.25             1 55.57596 -19.57596491
#> 14               39 1948.50             1 42.87393  -3.87392646
#> 15               NA 1948.75             1       NA           NA
#> 16               NA 1949.00             0       NA           NA
#> 17               69 1949.25             0 52.67573  16.32427049
#> 18               57 1949.50             0 65.52282  -8.52282424
#> 19               57 1949.75             0 59.05786  -2.05785524
#> 20               51 1950.00             0 57.35395  -6.35394558
#> 21               45 1950.25             0 53.29794  -8.29793582
#> 22               37 1950.50             0 48.38997 -11.38997122
#> 23               46 1950.75             0 42.13000   3.86999663
#> 24               39 1951.00             0 47.07808  -8.07807825
#> 25               36 1951.25             0 43.62400  -7.62399910
#> 26               24 1951.50             0 40.60205 -16.60204691
#> 27               32 1951.75             0 32.06405  -0.06404997
#> 28               23 1952.00             1 35.76815 -12.76815333
#> 29               25 1952.25             1 30.82008  -5.82007845
#> 30               32 1952.50             1 30.89415   1.10585053
#> 31               NA 1952.75             1       NA           NA
#> 32               32 1953.00             0 38.25271  -6.25271200
#> 33               59 1953.25             0 37.11299  21.88700556
#> 34               74 1953.50             0 55.15614  18.84386295
#> 35               75 1953.75             0 69.12996   5.87004180
#> 36               60 1954.00             0 71.93585 -11.93584690
#> 37               71 1954.25             0 61.93781   9.06218505
#> 38               61 1954.50             0 67.24395  -6.24394579
#> 39               71 1954.75             0 62.04585   8.95415331
#> 40               57 1955.00             0 67.38594 -10.38593826
#> 41               71 1955.25             0 59.34184  11.65815982
#> 42               68 1955.50             0 66.81797   1.18203163
#> 43               79 1955.75             0 66.77786  12.22214191
#> 44               73 1956.00             1 73.78790  -0.78789858
#> 45               76 1956.25             1 71.29381   4.70619400
#> 46               71 1956.50             1 72.46986  -1.46985605
#> 47               67 1956.75             1 69.51583  -2.51582533
#> 48               75 1957.00             0 66.10186   8.89814354
#> 49               79 1957.25             0 70.94190   8.05810040
#> 50               62 1957.50             0 74.78185 -12.78184588
#> 51               63 1957.75             0 63.85779  -0.85778809
#> 52               57 1958.00             0 62.11992  -5.11991771
#> 53               60 1958.25             0 58.20590   1.79409959
#> 54               49 1958.50             0 59.38195 -10.38195047
#> 55               48 1958.75             0 52.37191  -4.37190997
#> 56               52 1959.00             0 50.13399   1.86600884
#> 57               57 1959.25             0 52.69601   4.30399480
#> 58               62 1959.50             0 56.64398   5.35601677
#> 59               61 1959.75             0 60.73395   0.26604628
#> 60               66 1960.00             1 60.76791   5.23208555
#> 61               71 1960.25             1 64.00593   6.99406988
#> 62               62 1960.50             1 68.09590  -6.09590061
#> 63               61 1960.75             1 62.72185  -1.72184832
#> 64               57 1961.00             0 60.76791  -3.76791445
#> 65               72 1961.25             0 57.92192  14.07808453
#> 66               83 1961.50             0 67.49397  15.50603000
#> 67               71 1961.75             0 77.05987  -6.05987498
#> 68               78 1962.00             0 70.50977   7.49022737
#> 69               79 1962.25             0 73.53787   5.46212563
#> 70               71 1962.50             0 75.20782  -4.20782329
#> 71               62 1962.75             0 69.94180  -7.94180274
#> 72               74 1963.00             0 62.72185  11.27815168
#> 73               76 1963.25             0 69.55594   6.44406439
#> 74               64 1963.50             0 72.61185  -8.61184852
#> 75               62 1963.75             0 64.78381  -2.78381393
#> 76               57 1964.00             1 61.72790  -4.72790102
#> 77               80 1964.25             1 58.06391  21.93609206
#> 78               73 1964.50             1 72.90198   0.09801698
#> 79               69 1964.75             1 71.43580  -2.43579847
#> 80               69 1965.00             0 67.73784   1.26215534
#> 81               71 1965.25             0 67.16987   3.83012523
#> 82               64 1965.50             0 68.52188  -4.52187803
#> 83               69 1965.75             0 64.07385   4.92614842
#> 84               62 1966.00             0 66.45991  -4.45991242
#> 85               63 1966.25             0 62.43786   0.56213662
#> 86               46 1966.50             0 62.11992 -16.11991771
#> 87               56 1966.75             0 50.76988   5.23011750
#> 88               44 1967.00             0 55.11603 -11.11602677
#> 89               44 1967.25             0 48.42393  -4.42393195
#> 90               52 1967.50             0 46.72002   5.27997771
#> 91               38 1967.75             0 52.12804 -14.12803532
#> 92               46 1968.00             1 43.79995   2.20004770
#> 93               36 1968.25             1 47.22007 -11.22007072
#> 94               49 1968.50             1 41.59599   7.40400579
#> 95               35 1968.75             1 48.96409 -13.96409066
#> 96               44 1969.00             0 41.34597   2.65403000
#> 97               59 1969.25             0 45.44209  13.55790995
#> 98               65 1969.50             0 56.86005   8.13995329
#> 99               65 1969.75             0 63.04594   1.95405645
#> 100              56 1970.00             0 63.89790  -7.89789838
#> 101              66 1970.25             0 57.81388   8.18611627
#> 102              53 1970.50             0 63.29597 -10.29596776
#> 103              61 1970.75             0 55.92787   5.07212869
#> 104              52 1971.00             0 59.48998  -7.48998221
#> 105              51 1971.25             0 54.54191  -3.54190733
#> 106              48 1971.50             0 52.58797  -4.58797346
#> 107              54 1971.75             0 50.41798   3.58202390
#> 108              49 1972.00             1 54.04801  -5.04800846
#> 109              49 1972.25             1 51.51996  -2.51995515
#> 110              61 1972.50             1 50.80999  10.19000721
#> 111              NA 1972.75             1 58.92201           NA
#> 112              NA 1973.00             0 59.22120           NA
#> 113              68 1973.25             0 59.12839   8.87160905
#> 114              44 1973.50             0 59.10813 -15.10813486
#> 115              40 1973.75             0 59.08126 -19.08126367
#> 116              27 1974.00             0 59.06022 -32.06022249
#> 117              28 1974.25             0 59.04218 -31.04218311
#> 118              25 1974.50             0 59.02700 -34.02700077
#> 119              24 1974.75             0 59.01418 -35.01417603
#> 120              24 1975.00             0 59.00335 -35.00335071

# We can supply an external predictor as well
fit <- stats::arima(presidents_df$approval_rating[1:110], order = c(2, 0, 0),
                    xreg = as.matrix(presidents_df[1:110, "election_year", drop = FALSE]))
extract_yhat(fit, presidents_df, resp_var = approval_rating,
             newxreg = as.matrix(presidents_df[, "election_year", drop = FALSE]))
#>     approval_rating    date election_year  .fitted      .resid
#> 1                NA 1945.25             0       NA          NA
#> 2                87 1945.50             0 70.43649  16.5635128
#> 3                82 1945.75             0 81.19555   0.8044465
#> 4                75 1946.00             0 78.79648  -3.7964778
#> 5                63 1946.25             0 73.37698 -10.3769849
#> 6                50 1946.50             0 64.33458 -14.3345791
#> 7                43 1946.75             0 53.84118 -10.8411782
#> 8                32 1947.00             0 47.15836 -15.1583593
#> 9                35 1947.25             0 38.77737  -3.7773698
#> 10               60 1947.50             0 39.02455  20.9754546
#> 11               54 1947.75             0 56.03370  -2.0337002
#> 12               55 1948.00             1 52.51064   2.4893584
#> 13               36 1948.25             1 54.54114 -18.5411439
#> 14               39 1948.50             1 42.68524  -3.6852427
#> 15               NA 1948.75             1       NA          NA
#> 16               NA 1949.00             0       NA          NA
#> 17               69 1949.25             0 54.29341  14.7065935
#> 18               57 1949.50             0 65.68901  -8.6890065
#> 19               57 1949.75             0 59.41859  -2.4185868
#> 20               51 1950.00             0 57.52360  -6.5235978
#> 21               45 1950.25             0 53.55510  -8.5551000
#> 22               37 1950.50             0 48.63911 -11.6391077
#> 23               46 1950.75             0 42.40028   3.5997172
#> 24               39 1951.00             0 47.08970  -8.0897035
#> 25               36 1951.25             0 43.88103  -7.8810311
#> 26               24 1951.50             0 40.79137 -16.7913720
#> 27               32 1951.75             0 32.38063  -0.3806291
#> 28               23 1952.00             1 32.27452  -9.2745160
#> 29               25 1952.25             1 29.90168  -4.9016758
#> 30               32 1952.50             1 30.35636   1.6436406
#> 31               NA 1952.75             1       NA          NA
#> 32               32 1953.00             0 40.41909  -8.4190913
#> 33               59 1953.25             0 37.56652  21.4334830
#> 34               74 1953.50             0 54.89854  19.1014634
#> 35               75 1953.75             0 69.08351   5.9164936
#> 36               60 1954.00             0 72.11366 -12.1136589
#> 37               71 1954.25             0 62.35033   8.6496698
#> 38               61 1954.50             0 67.25717  -6.2571732
#> 39               71 1954.75             0 62.38008   8.6199165
#> 40               57 1955.00             0 67.41509 -10.4150890
#> 41               71 1955.25             0 59.73442  11.2655817
#> 42               68 1955.50             0 66.78343   1.2165740
#> 43               79 1955.75             0 67.01000  11.9900024
#> 44               73 1956.00             1 70.30938   2.6906249
#> 45               76 1956.25             1 70.39453   5.6054690
#> 46               71 1956.50             1 71.98438  -0.9843782
#> 47               67 1956.75             1 69.15104  -2.1510439
#> 48               75 1957.00             0 69.21825   5.7817455
#> 49               79 1957.25             0 71.56134   7.4386586
#> 50               62 1957.50             0 74.91724 -12.9172399
#> 51               63 1957.75             0 64.30483  -1.3048258
#> 52               57 1958.00             0 62.28167  -5.2816743
#> 53               60 1958.25             0 58.47109   1.5289077
#> 54               49 1958.50             0 59.50785 -10.5078467
#> 55               48 1958.75             0 52.70601  -4.7060146
#> 56               52 1959.00             0 50.30753   1.6924749
#> 57               57 1959.25             0 52.79527   4.2047255
#> 58               62 1959.50             0 56.73402   5.2659810
#> 59               61 1959.75             0 60.83068   0.1693207
#> 60               66 1960.00             1 57.45639   8.5436128
#> 61               71 1960.25             1 62.92213   8.0778666
#> 62               62 1960.50             1 67.57189  -5.5718864
#> 63               61 1960.75             1 62.40872  -1.4087185
#> 64               57 1961.00             0 63.82851  -6.8285150
#> 65               72 1961.25             0 58.70835  13.2916465
#> 66               83 1961.50             0 67.44484  15.5551577
#> 67               71 1961.75             0 77.08916  -6.0891578
#> 68               78 1962.00             0 70.88924   7.1107645
#> 69               79 1962.25             0 73.62416   5.3758394
#> 70               71 1962.50             0 75.39099  -4.3909871
#> 71               62 1962.75             0 70.25757  -8.2575725
#> 72               74 1963.00             0 63.04150  10.9585002
#> 73               76 1963.25             0 69.55725   6.4427464
#> 74               64 1963.50             0 72.77508  -8.7750752
#> 75               62 1963.75             0 65.15391  -3.1539111
#> 76               57 1964.00             1 58.43363  -1.4336350
#> 77               80 1964.25             1 57.12730  22.8726975
#> 78               73 1964.50             1 72.10339   0.8966086
#> 79               69 1964.75             1 71.10554  -2.1055395
#> 80               69 1965.00             0 70.85692  -1.8569186
#> 81               71 1965.25             0 67.90868   3.0913249
#> 82               64 1965.50             0 68.67841  -4.6784150
#> 83               69 1965.75             0 64.36433   4.6356676
#> 84               62 1966.00             0 66.56600  -4.5660036
#> 85               63 1966.25             0 62.72567   0.2743317
#> 86               46 1966.50             0 62.28167 -16.2816743
#> 87               56 1966.75             0 51.19551   4.8044870
#> 88               44 1967.00             0 55.12511 -11.1251082
#> 89               44 1967.25             0 48.76727  -4.7672701
#> 90               52 1967.50             0 46.87228   5.1277189
#> 91               38 1967.75             0 52.16361 -14.1636115
#> 92               46 1968.00             1 40.66465   5.3353452
#> 93               36 1968.25             1 46.06175 -10.0617452
#> 94               49 1968.50             1 41.26400   7.7359991
#> 95               35 1968.75             1 48.28326 -13.2832553
#> 96               44 1969.00             0 44.57879  -0.5787864
#> 97               59 1969.25             0 46.00413  12.9958679
#> 98               65 1969.50             0 56.79353   8.2064744
#> 99               65 1969.75             0 63.13076   1.8692403
#> 100              56 1970.00             0 64.07825  -8.0782542
#> 101              66 1970.25             0 58.12551   7.8744925
#> 102              53 1970.50             0 63.31843 -10.3184287
#> 103              61 1970.75             0 56.29917   4.7008257
#> 104              52 1971.00             0 59.53760  -7.5376000
#> 105              51 1971.25             0 54.84818  -3.8481793
#> 106              48 1971.50             0 52.76552  -4.7655212
#> 107              54 1971.75             0 50.62336   3.3766434
#> 108              49 1972.00             1 50.61565  -1.6156526
#> 109              49 1972.25             1 50.57265  -1.5726461
#> 110              61 1972.50             1 50.33616  10.6638399
#> 111              NA 1972.75             1 58.27316          NA
#> 112              NA 1973.00             0 61.86702          NA
#> 113              68 1973.25             0 61.49687   6.5031319
#> 114              44 1973.50             0 61.26648 -17.2664787
#> 115              40 1973.75             0 61.05564 -21.0556426
#> 116              27 1974.00             0 60.87981 -33.8798101
#> 117              28 1974.25             0 60.73022 -32.7302173
#> 118              25 1974.50             0 60.60351 -35.6035074
#> 119              24 1974.75             0 60.49608 -36.4960763
#> 120              24 1975.00             0 60.40501 -36.4050102

# We can use arima_tidy to not have to generate xreg manually
fit <- arima_tidy(approval_rating ~ election_year, order = c(2, 0, 0),
                  data = presidents_df[1:110, ])
extract_yhat(fit, presidents_df)
#>     approval_rating    date election_year  .fitted      .resid
#> 1                NA 1945.25             0       NA          NA
#> 2                87 1945.50             0 70.43649  16.5635128
#> 3                82 1945.75             0 81.19555   0.8044465
#> 4                75 1946.00             0 78.79648  -3.7964778
#> 5                63 1946.25             0 73.37698 -10.3769849
#> 6                50 1946.50             0 64.33458 -14.3345791
#> 7                43 1946.75             0 53.84118 -10.8411782
#> 8                32 1947.00             0 47.15836 -15.1583593
#> 9                35 1947.25             0 38.77737  -3.7773698
#> 10               60 1947.50             0 39.02455  20.9754546
#> 11               54 1947.75             0 56.03370  -2.0337002
#> 12               55 1948.00             1 52.51064   2.4893584
#> 13               36 1948.25             1 54.54114 -18.5411439
#> 14               39 1948.50             1 42.68524  -3.6852427
#> 15               NA 1948.75             1       NA          NA
#> 16               NA 1949.00             0       NA          NA
#> 17               69 1949.25             0 54.29341  14.7065935
#> 18               57 1949.50             0 65.68901  -8.6890065
#> 19               57 1949.75             0 59.41859  -2.4185868
#> 20               51 1950.00             0 57.52360  -6.5235978
#> 21               45 1950.25             0 53.55510  -8.5551000
#> 22               37 1950.50             0 48.63911 -11.6391077
#> 23               46 1950.75             0 42.40028   3.5997172
#> 24               39 1951.00             0 47.08970  -8.0897035
#> 25               36 1951.25             0 43.88103  -7.8810311
#> 26               24 1951.50             0 40.79137 -16.7913720
#> 27               32 1951.75             0 32.38063  -0.3806291
#> 28               23 1952.00             1 32.27452  -9.2745160
#> 29               25 1952.25             1 29.90168  -4.9016758
#> 30               32 1952.50             1 30.35636   1.6436406
#> 31               NA 1952.75             1       NA          NA
#> 32               32 1953.00             0 40.41909  -8.4190913
#> 33               59 1953.25             0 37.56652  21.4334830
#> 34               74 1953.50             0 54.89854  19.1014634
#> 35               75 1953.75             0 69.08351   5.9164936
#> 36               60 1954.00             0 72.11366 -12.1136589
#> 37               71 1954.25             0 62.35033   8.6496698
#> 38               61 1954.50             0 67.25717  -6.2571732
#> 39               71 1954.75             0 62.38008   8.6199165
#> 40               57 1955.00             0 67.41509 -10.4150890
#> 41               71 1955.25             0 59.73442  11.2655817
#> 42               68 1955.50             0 66.78343   1.2165740
#> 43               79 1955.75             0 67.01000  11.9900024
#> 44               73 1956.00             1 70.30938   2.6906249
#> 45               76 1956.25             1 70.39453   5.6054690
#> 46               71 1956.50             1 71.98438  -0.9843782
#> 47               67 1956.75             1 69.15104  -2.1510439
#> 48               75 1957.00             0 69.21825   5.7817455
#> 49               79 1957.25             0 71.56134   7.4386586
#> 50               62 1957.50             0 74.91724 -12.9172399
#> 51               63 1957.75             0 64.30483  -1.3048258
#> 52               57 1958.00             0 62.28167  -5.2816743
#> 53               60 1958.25             0 58.47109   1.5289077
#> 54               49 1958.50             0 59.50785 -10.5078467
#> 55               48 1958.75             0 52.70601  -4.7060146
#> 56               52 1959.00             0 50.30753   1.6924749
#> 57               57 1959.25             0 52.79527   4.2047255
#> 58               62 1959.50             0 56.73402   5.2659810
#> 59               61 1959.75             0 60.83068   0.1693207
#> 60               66 1960.00             1 57.45639   8.5436128
#> 61               71 1960.25             1 62.92213   8.0778666
#> 62               62 1960.50             1 67.57189  -5.5718864
#> 63               61 1960.75             1 62.40872  -1.4087185
#> 64               57 1961.00             0 63.82851  -6.8285150
#> 65               72 1961.25             0 58.70835  13.2916465
#> 66               83 1961.50             0 67.44484  15.5551577
#> 67               71 1961.75             0 77.08916  -6.0891578
#> 68               78 1962.00             0 70.88924   7.1107645
#> 69               79 1962.25             0 73.62416   5.3758394
#> 70               71 1962.50             0 75.39099  -4.3909871
#> 71               62 1962.75             0 70.25757  -8.2575725
#> 72               74 1963.00             0 63.04150  10.9585002
#> 73               76 1963.25             0 69.55725   6.4427464
#> 74               64 1963.50             0 72.77508  -8.7750752
#> 75               62 1963.75             0 65.15391  -3.1539111
#> 76               57 1964.00             1 58.43363  -1.4336350
#> 77               80 1964.25             1 57.12730  22.8726975
#> 78               73 1964.50             1 72.10339   0.8966086
#> 79               69 1964.75             1 71.10554  -2.1055395
#> 80               69 1965.00             0 70.85692  -1.8569186
#> 81               71 1965.25             0 67.90868   3.0913249
#> 82               64 1965.50             0 68.67841  -4.6784150
#> 83               69 1965.75             0 64.36433   4.6356676
#> 84               62 1966.00             0 66.56600  -4.5660036
#> 85               63 1966.25             0 62.72567   0.2743317
#> 86               46 1966.50             0 62.28167 -16.2816743
#> 87               56 1966.75             0 51.19551   4.8044870
#> 88               44 1967.00             0 55.12511 -11.1251082
#> 89               44 1967.25             0 48.76727  -4.7672701
#> 90               52 1967.50             0 46.87228   5.1277189
#> 91               38 1967.75             0 52.16361 -14.1636115
#> 92               46 1968.00             1 40.66465   5.3353452
#> 93               36 1968.25             1 46.06175 -10.0617452
#> 94               49 1968.50             1 41.26400   7.7359991
#> 95               35 1968.75             1 48.28326 -13.2832553
#> 96               44 1969.00             0 44.57879  -0.5787864
#> 97               59 1969.25             0 46.00413  12.9958679
#> 98               65 1969.50             0 56.79353   8.2064744
#> 99               65 1969.75             0 63.13076   1.8692403
#> 100              56 1970.00             0 64.07825  -8.0782542
#> 101              66 1970.25             0 58.12551   7.8744925
#> 102              53 1970.50             0 63.31843 -10.3184287
#> 103              61 1970.75             0 56.29917   4.7008257
#> 104              52 1971.00             0 59.53760  -7.5376000
#> 105              51 1971.25             0 54.84818  -3.8481793
#> 106              48 1971.50             0 52.76552  -4.7655212
#> 107              54 1971.75             0 50.62336   3.3766434
#> 108              49 1972.00             1 50.61565  -1.6156526
#> 109              49 1972.25             1 50.57265  -1.5726461
#> 110              61 1972.50             1 50.33616  10.6638399
#> 111              NA 1972.75             1 58.27316          NA
#> 112              NA 1973.00             0 61.86702          NA
#> 113              68 1973.25             0 61.49687   6.5031319
#> 114              44 1973.50             0 61.26648 -17.2664787
#> 115              40 1973.75             0 61.05564 -21.0556426
#> 116              27 1974.00             0 60.87981 -33.8798101
#> 117              28 1974.25             0 60.73022 -32.7302173
#> 118              25 1974.50             0 60.60351 -35.6035074
#> 119              24 1974.75             0 60.49608 -36.4960763
#> 120              24 1975.00             0 60.40501 -36.4050102