Calculate the Gaussian CUSUM statistic over multiple spatial regions

parallel_cusum_gaussian calculates the Gaussian CUSUM statistic independently over each spatial region, over all time points. Unlike e.g., scan_cusum_poisson, information from multiple spatial regions is never aggregated.

Usage

parallel_cusum_gaussian(
  wide_cases,
  wide_baseline = NULL,
  mean = 0,
  sigma = 1,
  drift = 0.5
)

Arguments

wide_cases

A matrix of observed counts. Rows indicate time and are ordered from least recent (row 1) to most recent (row nrow(counts)). Columns indicate locations, numbered from 1 and up. Can be generated from a data.frame with the pivot_for_scan function.

wide_baseline

A matrix of the same dimensions as wide_cases, holding the expected value of cases at each space-time location.This value is typically estimated from past data using a GLM or other model.

mean

The estimated mean of the residuals (wide_cases - wide_baseline)

sigma

The estimated standard deviation of the residuals.

drift

The downward drift to use in calculating the CUSUM statistic. The value of drift is typically half of the expected increase in the value of the standardized residuals during an outbreak.

Value

A matrix with the same dimensions as wide_cases, where rows indicate time (oldest to newest) and columns indicate locations. Element (i, j) indicates the value of the CUSUM statistic at time i in region j. The value of the CUSUM statistic at row i incorporates information from time points 1 thought i.

Details

The CUSUM statistic is widely used surveillance technique for purely temporal processes. It is calculated by taking the cumulative sum of the standard residuals with a downward drift and a floor of 0. For a Gaussian process, residuals are calculated as resid = wide_cases - wide_baseline, and standard residuals are calculated as (resid - mean)/sigma. The CUSUM statistic is then calculated by summing the standardized residuals with downward drift drift.

The three parameters used in calculating the standardized residuals and CUSUM statistic, mean, sigma, and drift, can each be a scalar, a vector with a separate value for each spatial region, or a matrix with the same dimensions as wide_cases

See also

Examples

outbreak_sp <- c(1, 2, 4, 5)
outbreak_tm <- c(3, 4)
wide_cases <- matrix(2, nrow = 4, ncol = 9)
wide_cases[outbreak_tm, outbreak_sp] <- 5
wide_cases[c(3, 4), c(8, 9)] <- 2.5
wide_baseline <- matrix(2, nrow = 4, ncol = 9)
scanres <- parallel_cusum_gaussian(wide_cases, wide_baseline,
                                   mean = 1,
                                   sigma = 2,
                                   drift = 0.5)

v_mean <- c(rep(0.0, 8), 2)
m_sigma <- matrix(1.5, nrow = 4, ncol = 9)
m_sigma[c(3, 4), c(4, 5)] <- 0.25
scanres <- parallel_cusum_gaussian(wide_cases, wide_baseline,
                                   mean = v_mean,
                                   sigma = m_sigma,
                                   drift = 0.5)