Turn a long data frame into a wide matrix

A friendly wrapper for pivot_wider that takes a long data.frame and transforms it into a wide matrix. This is the format expected as input for the #' scanstatistics package.

Usage

pivot_for_scan(spacetime_data, value_col)

Arguments

spacetime_data

A data frame or a data frame extension (e.g., a tibble) that includes columns `id_space` and `id_time` that are each a set of consecutive integers. The column `id_time` identifies the time point associated with each row, with lower numbers being earlier and higher numbers being later. The column `id_space` identifies the spatial location or area associated with each row, but no meaning or order is ascribed to the values.

value_col

An expression defining the values in in spacetime_data to be pivoted into the wide matrix. This argument uses data masking and should be specified as bare code (not a string). Can also be a scalar, in which case all entries in the wide matrix will be equal to value_col.

Value

A matrix, with rows corresponding to values of id_time and columns corresponding to id_space. Exception: if value_col = NULL, then returns NULL.

Details

Functions for calculating scan statistics in the scanstatistics package expect case counts and baseline estimates to be formatted as matrices, where the rows correspond to time (organized from earliest to latest) and the columns correspond to space.

pivot_for_scan takes spacetime_data, and pivots the data into that format. The rows are ordered by the id_time column from smaller to largest, and the columns are ordered by the id_space column in spacetime_data, also from smallest to largest. pivot_for_scan expects the data to be complete, so that every combination of id_space and id_time is associated with exactly one value. If some combination of id_space and id_time has no corresponding value in spacetime_data, then that entry in the returned matrix will be <NA>.

Examples

df <- data.frame(vals = c(1, 2, 3, 4, 5, 6),
                 id_space = rep(1:3, 2),
                 id_time = rep(1:2, each = 3))

pivot_for_scan(df, vals)
#>   1 2 3
#> 1 1 2 3
#> 2 4 5 6

# Non-exist combinations create NA's
incomplete_df <- df[1:5, ]
pivot_for_scan(incomplete_df, vals)
#>   1 2  3
#> 1 1 2  3
#> 2 4 5 NA

# Returns NULL:
pivot_for_scan(df, NULL)
#> NULL