Forecast from arima_tidy fits

Forecast from models fit by arima_tidy. Calls predict.Arima under the hood.

Usage

# S3 method for arima_tidy
predict(object, only_newdata = NULL, n.ahead, se.fit = TRUE, ...)

Arguments

object

The result of an arima_tidy fit

only_newdata

New data to be used for prediction. Will be transformed into newxreg using the formula stored in object.

n.ahead

The number of steps ahead for which prediction is required. Optional if only_newdata is specified, otherwise required.

se.fit

Logical: should standard errors of prediction be returned?

...

arguments passed to or from other methods.

Value

A time series of predictions, or if se.fit = TRUE, a list with timeseries pred and se.

Details

Even though arima_tidy models are fit like lm or glm models, it's important to remember that they behave differently, because the order of the data points is crucial for all ARIMA models. The rows in only_newdata are assumed to be ordered from earliest to latest, and are assumed to follow immeadiately after the data used to fit the model. To extract fitted values and generate predictions in a single step, use augment.arima_tidy.

See also

Examples

data(presidents_df)
# Fit on the first 110 rows, predict the last 10
fit <- arima_tidy(approval_rating ~ 1,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
predict(fit, n.ahead = 10)
#> $pred
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>  [1] 60.50137 60.10897 59.80017 59.55715 59.36591 59.21541 59.09698 59.00377
#>  [9] 58.93042 58.87270
#> 
#> $se
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>  [1]  9.082644 11.557828 12.853905 13.594751 14.033962 14.299204 14.461029
#>  [8] 14.560346 14.621515 14.659269
#> 

fit <- arima_tidy(approval_rating ~ election_year,
                  data = presidents_df[1:110, ],
                  order = c(1, 0, 0))
predict(fit, only_newdata = presidents_df[111:120, ], n.ahead = 10)
#> $pred
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>      111      112      113      114      115      116      117      118 
#> 59.99809 62.32400 61.70657 61.22189 60.84140 60.54271 60.30824 60.12417 
#>      119      120 
#> 59.97968 59.86625 
#> 
#> $se
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>  [1]  9.031329 11.481687 12.759482 13.486764 13.916035 14.174099 14.330817
#>  [8] 14.426546 14.485224 14.521266
#> 
predict(fit, only_newdata = presidents_df[111:120, ])
#> $pred
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>      111      112      113      114      115      116      117      118 
#> 59.99809 62.32400 61.70657 61.22189 60.84140 60.54271 60.30824 60.12417 
#>      119      120 
#> 59.97968 59.86625 
#> 
#> $se
#> Time Series:
#> Start = 111 
#> End = 120 
#> Frequency = 1 
#>  [1]  9.031329 11.481687 12.759482 13.486764 13.916035 14.174099 14.330817
#>  [8] 14.426546 14.485224 14.521266
#>