Prepare data for prediction in statistical surveillance

Prepare data for prediction in statistical surveillance

Usage

prepare_training_data(
  spacetime_data,
  outcome_cols,
  split_id,
  prep_strategy = c("NA", "truncate")
)

Arguments

spacetime_data

A data frame or a data frame extension (e.g., a tibble) that includes columns `id_space` and `id_time` that are each a set of consecutive integers. The column `id_time` identifies the time point associated with each row, with lower numbers being earlier and higher numbers being later. The column `id_space` identifies the spatial location or area associated with each row, but no meaning or order is ascribed to the values.

outcome_cols

<tidyr_tidy_select> Outcome columns that should be removed or NA'd out of the training data. Follows the tidy-select principles of the tidyverse.

split_id

The value of id_time that marks the beginning of the prediction data. All entries in the outcome columns associated with an id_time that is greater than or equal to split_id will be removed or NA'd out of the training data.

prep_strategy

Whether the values should be converted to NA (prep_strategy = "NA") or if those rows should be removed (prep_strategy = "truncate")

Value

A data frame with the selected values converted to NA or removed, as designated by prep_strategy

Examples

x = rnorm(100)
spacetime_data <- data.frame(id_space = rep(1:10, each = 10),
                             id_time = rep(1:10, 10),
                             x = x,
                             y = 2.04 * x + 1.23,
                             y2 = (2.04 * x + 1.23)^2)

# NA-out a single column
prepare_training_data(spacetime_data, y, split_id = 9)
#>     id_space id_time            x           y           y2
#> 1          1       1  0.846869397  2.95761357 8.747478e+00
#> 2          1       2  0.289797103  1.82118609 3.316719e+00
#> 3          1       3 -0.455661138  0.30045128 9.027097e-02
#> 4          1       4 -0.120609164  0.98395730 9.681720e-01
#> 5          1       5  0.323701479  1.89035102 3.573427e+00
#> 6          1       6 -0.242792754  0.73470278 5.397882e-01
#> 7          1       7 -1.162309122 -1.14111061 1.302133e+00
#> 8          1       8 -1.628357761 -2.09184983 4.375836e+00
#> 9          1       9  1.099096129          NA 1.205587e+01
#> 10         1      10  0.095564740          NA 2.030488e+00
#> 11         2       1 -1.192322471 -1.20233784 1.445616e+00
#> 12         2       2 -0.292184020  0.63394460 4.018858e-01
#> 13         2       3  0.904031481  3.07422422 9.450855e+00
#> 14         2       4  0.378723623  2.00259619 4.010392e+00
#> 15         2       5  0.604929568  2.46405632 6.071574e+00
#> 16         2       6 -1.000137395 -0.81028029 6.565541e-01
#> 17         2       7  0.807282402  2.87685610 8.276301e+00
#> 18         2       8 -0.376117252  0.46272081 2.141105e-01
#> 19         2       9 -0.696751935          NA 3.662399e-02
#> 20         2      10 -0.673842044          NA 2.092008e-02
#> 21         3       1  0.798776048  2.85950314 8.176758e+00
#> 22         3       2  0.597404118  2.44870440 5.996153e+00
#> 23         3       3 -0.311625673  0.59428363 3.531730e-01
#> 24         3       4 -0.393228151  0.42781457 1.830253e-01
#> 25         3       5  2.188897121  5.69535013 3.243701e+01
#> 26         3       6  0.461360015  2.17117443 4.713998e+00
#> 27         3       7  0.974930068  3.21885734 1.036104e+01
#> 28         3       8  0.485507540  2.22043538 4.930333e+00
#> 29         3       9  1.448336223          NA 1.751093e+01
#> 30         3      10  1.119655519          NA 1.234888e+01
#> 31         4       1 -0.805235888 -0.41268121 1.703058e-01
#> 32         4       2  1.600780237  4.49559168 2.021034e+01
#> 33         4       3 -0.715547221 -0.22971633 5.276959e-02
#> 34         4       4 -1.287071150 -1.39562515 1.947770e+00
#> 35         4       5  0.815259874  2.89313014 8.370202e+00
#> 36         4       6 -0.070630992  1.08591278 1.179207e+00
#> 37         4       7 -0.018097109  1.19308190 1.423444e+00
#> 38         4       8 -0.970277248 -0.74936559 5.615488e-01
#> 39         4       9 -0.046794614          NA 1.287179e+00
#> 40         4      10  0.876179522          NA 9.104740e+00
#> 41         5       1 -0.045629610  1.13691559 1.292577e+00
#> 42         5       2  0.591829528  2.43733224 5.940588e+00
#> 43         5       3  0.390773389  2.02717771 4.109449e+00
#> 44         5       4  0.335713472  1.91485548 3.666672e+00
#> 45         5       5  1.356254073  3.99675831 1.597408e+01
#> 46         5       6 -0.214746801  0.79191653 6.271318e-01
#> 47         5       7 -0.012695749  1.20410067 1.449858e+00
#> 48         5       8 -0.928024047 -0.66316906 4.397932e-01
#> 49         5       9  1.953132106          NA 2.718986e+01
#> 50         5      10  1.492505140          NA 1.827315e+01
#> 51         6       1 -0.595075878  0.01604521 2.574487e-04
#> 52         6       2 -1.633478913 -2.10229698 4.419653e+00
#> 53         6       3  1.332074224  3.94743142 1.558221e+01
#> 54         6       4  0.186611837  1.61068815 2.594316e+00
#> 55         6       5 -0.008543471  1.21257132 1.470329e+00
#> 56         6       6  0.497079901  2.24404300 5.035729e+00
#> 57         6       7  0.083847694  1.40104930 1.962939e+00
#> 58         6       8  0.326840468  1.89675455 3.597678e+00
#> 59         6       9  0.098897888          NA 2.049913e+00
#> 60         6      10  0.348056776          NA 3.763739e+00
#> 61         7       1 -0.103945433  1.01795132 1.036225e+00
#> 62         7       2  0.227346734  1.69378734 2.868916e+00
#> 63         7       3  2.031012081  5.37326465 2.887197e+01
#> 64         7       4  0.221963908  1.68280637 2.831837e+00
#> 65         7       5  2.637643737  6.61079322 4.370259e+01
#> 66         7       6 -1.443755569 -1.71526136 2.942122e+00
#> 67         7       7 -0.956117286 -0.72047926 5.190904e-01
#> 68         7       8 -0.182490946  0.85771847 7.356810e-01
#> 69         7       9  1.861330816          NA 2.527188e+01
#> 70         7      10  0.828443485          NA 8.526544e+00
#> 71         8       1  0.531210254  2.31366892 5.353064e+00
#> 72         8       2 -0.004128931  1.22157698 1.492250e+00
#> 73         8       3 -0.676207567 -0.14946344 2.233932e-02
#> 74         8       4  0.228571453  1.69628576 2.877385e+00
#> 75         8       5  2.060793529  5.43401880 2.952856e+01
#> 76         8       6  1.250625379  3.78127577 1.429805e+01
#> 77         8       7  0.394004558  2.03376930 4.136218e+00
#> 78         8       8  0.542711034  2.33713051 5.462179e+00
#> 79         8       9 -0.305519502          NA 3.681337e-01
#> 80         8      10  2.508846475          NA 4.029770e+01
#> 81         9       1  0.713624674  2.68579434 7.213491e+00
#> 82         9       2  1.176272934  3.62959678 1.317397e+01
#> 83         9       3  1.220099799  3.71900359 1.383099e+01
#> 84         9       4 -0.701922247 -0.20192138 4.077225e-02
#> 85         9       5  0.498580322  2.24710386 5.049476e+00
#> 86         9       6  0.334212564  1.91179363 3.654955e+00
#> 87         9       7  1.082418318  3.43813337 1.182076e+01
#> 88         9       8 -1.268606521 -1.35795730 1.844048e+00
#> 89         9       9 -0.604883499          NA 1.570012e-05
#> 90         9      10  0.242484199          NA 2.974479e+00
#> 91        10       1  1.766816262  4.83430517 2.337051e+01
#> 92        10       2 -2.061966415 -2.97641149 8.859025e+00
#> 93        10       3 -0.991669740 -0.79300627 6.288589e-01
#> 94        10       4 -0.978416117 -0.76596888 5.867083e-01
#> 95        10       5 -0.209345980  0.80293420 6.447033e-01
#> 96        10       6  1.444200333  4.17616868 1.744038e+01
#> 97        10       7  0.144674259  1.52513549 2.326038e+00
#> 98        10       8 -2.126999077 -3.10907812 9.666367e+00
#> 99        10       9  0.464926452          NA 4.745644e+00
#> 100       10      10  0.023903480          NA 1.635235e+00

# NA-out multiple columns
prepare_training_data(spacetime_data, c(y, y2), split_id = 9)
#>     id_space id_time            x           y           y2
#> 1          1       1  0.846869397  2.95761357 8.747478e+00
#> 2          1       2  0.289797103  1.82118609 3.316719e+00
#> 3          1       3 -0.455661138  0.30045128 9.027097e-02
#> 4          1       4 -0.120609164  0.98395730 9.681720e-01
#> 5          1       5  0.323701479  1.89035102 3.573427e+00
#> 6          1       6 -0.242792754  0.73470278 5.397882e-01
#> 7          1       7 -1.162309122 -1.14111061 1.302133e+00
#> 8          1       8 -1.628357761 -2.09184983 4.375836e+00
#> 9          1       9  1.099096129          NA           NA
#> 10         1      10  0.095564740          NA           NA
#> 11         2       1 -1.192322471 -1.20233784 1.445616e+00
#> 12         2       2 -0.292184020  0.63394460 4.018858e-01
#> 13         2       3  0.904031481  3.07422422 9.450855e+00
#> 14         2       4  0.378723623  2.00259619 4.010392e+00
#> 15         2       5  0.604929568  2.46405632 6.071574e+00
#> 16         2       6 -1.000137395 -0.81028029 6.565541e-01
#> 17         2       7  0.807282402  2.87685610 8.276301e+00
#> 18         2       8 -0.376117252  0.46272081 2.141105e-01
#> 19         2       9 -0.696751935          NA           NA
#> 20         2      10 -0.673842044          NA           NA
#> 21         3       1  0.798776048  2.85950314 8.176758e+00
#> 22         3       2  0.597404118  2.44870440 5.996153e+00
#> 23         3       3 -0.311625673  0.59428363 3.531730e-01
#> 24         3       4 -0.393228151  0.42781457 1.830253e-01
#> 25         3       5  2.188897121  5.69535013 3.243701e+01
#> 26         3       6  0.461360015  2.17117443 4.713998e+00
#> 27         3       7  0.974930068  3.21885734 1.036104e+01
#> 28         3       8  0.485507540  2.22043538 4.930333e+00
#> 29         3       9  1.448336223          NA           NA
#> 30         3      10  1.119655519          NA           NA
#> 31         4       1 -0.805235888 -0.41268121 1.703058e-01
#> 32         4       2  1.600780237  4.49559168 2.021034e+01
#> 33         4       3 -0.715547221 -0.22971633 5.276959e-02
#> 34         4       4 -1.287071150 -1.39562515 1.947770e+00
#> 35         4       5  0.815259874  2.89313014 8.370202e+00
#> 36         4       6 -0.070630992  1.08591278 1.179207e+00
#> 37         4       7 -0.018097109  1.19308190 1.423444e+00
#> 38         4       8 -0.970277248 -0.74936559 5.615488e-01
#> 39         4       9 -0.046794614          NA           NA
#> 40         4      10  0.876179522          NA           NA
#> 41         5       1 -0.045629610  1.13691559 1.292577e+00
#> 42         5       2  0.591829528  2.43733224 5.940588e+00
#> 43         5       3  0.390773389  2.02717771 4.109449e+00
#> 44         5       4  0.335713472  1.91485548 3.666672e+00
#> 45         5       5  1.356254073  3.99675831 1.597408e+01
#> 46         5       6 -0.214746801  0.79191653 6.271318e-01
#> 47         5       7 -0.012695749  1.20410067 1.449858e+00
#> 48         5       8 -0.928024047 -0.66316906 4.397932e-01
#> 49         5       9  1.953132106          NA           NA
#> 50         5      10  1.492505140          NA           NA
#> 51         6       1 -0.595075878  0.01604521 2.574487e-04
#> 52         6       2 -1.633478913 -2.10229698 4.419653e+00
#> 53         6       3  1.332074224  3.94743142 1.558221e+01
#> 54         6       4  0.186611837  1.61068815 2.594316e+00
#> 55         6       5 -0.008543471  1.21257132 1.470329e+00
#> 56         6       6  0.497079901  2.24404300 5.035729e+00
#> 57         6       7  0.083847694  1.40104930 1.962939e+00
#> 58         6       8  0.326840468  1.89675455 3.597678e+00
#> 59         6       9  0.098897888          NA           NA
#> 60         6      10  0.348056776          NA           NA
#> 61         7       1 -0.103945433  1.01795132 1.036225e+00
#> 62         7       2  0.227346734  1.69378734 2.868916e+00
#> 63         7       3  2.031012081  5.37326465 2.887197e+01
#> 64         7       4  0.221963908  1.68280637 2.831837e+00
#> 65         7       5  2.637643737  6.61079322 4.370259e+01
#> 66         7       6 -1.443755569 -1.71526136 2.942122e+00
#> 67         7       7 -0.956117286 -0.72047926 5.190904e-01
#> 68         7       8 -0.182490946  0.85771847 7.356810e-01
#> 69         7       9  1.861330816          NA           NA
#> 70         7      10  0.828443485          NA           NA
#> 71         8       1  0.531210254  2.31366892 5.353064e+00
#> 72         8       2 -0.004128931  1.22157698 1.492250e+00
#> 73         8       3 -0.676207567 -0.14946344 2.233932e-02
#> 74         8       4  0.228571453  1.69628576 2.877385e+00
#> 75         8       5  2.060793529  5.43401880 2.952856e+01
#> 76         8       6  1.250625379  3.78127577 1.429805e+01
#> 77         8       7  0.394004558  2.03376930 4.136218e+00
#> 78         8       8  0.542711034  2.33713051 5.462179e+00
#> 79         8       9 -0.305519502          NA           NA
#> 80         8      10  2.508846475          NA           NA
#> 81         9       1  0.713624674  2.68579434 7.213491e+00
#> 82         9       2  1.176272934  3.62959678 1.317397e+01
#> 83         9       3  1.220099799  3.71900359 1.383099e+01
#> 84         9       4 -0.701922247 -0.20192138 4.077225e-02
#> 85         9       5  0.498580322  2.24710386 5.049476e+00
#> 86         9       6  0.334212564  1.91179363 3.654955e+00
#> 87         9       7  1.082418318  3.43813337 1.182076e+01
#> 88         9       8 -1.268606521 -1.35795730 1.844048e+00
#> 89         9       9 -0.604883499          NA           NA
#> 90         9      10  0.242484199          NA           NA
#> 91        10       1  1.766816262  4.83430517 2.337051e+01
#> 92        10       2 -2.061966415 -2.97641149 8.859025e+00
#> 93        10       3 -0.991669740 -0.79300627 6.288589e-01
#> 94        10       4 -0.978416117 -0.76596888 5.867083e-01
#> 95        10       5 -0.209345980  0.80293420 6.447033e-01
#> 96        10       6  1.444200333  4.17616868 1.744038e+01
#> 97        10       7  0.144674259  1.52513549 2.326038e+00
#> 98        10       8 -2.126999077 -3.10907812 9.666367e+00
#> 99        10       9  0.464926452          NA           NA
#> 100       10      10  0.023903480          NA           NA

# We can use dplyr selection helpers
prepare_training_data(spacetime_data, dplyr::starts_with("y"), split_id = 9)
#>     id_space id_time            x           y           y2
#> 1          1       1  0.846869397  2.95761357 8.747478e+00
#> 2          1       2  0.289797103  1.82118609 3.316719e+00
#> 3          1       3 -0.455661138  0.30045128 9.027097e-02
#> 4          1       4 -0.120609164  0.98395730 9.681720e-01
#> 5          1       5  0.323701479  1.89035102 3.573427e+00
#> 6          1       6 -0.242792754  0.73470278 5.397882e-01
#> 7          1       7 -1.162309122 -1.14111061 1.302133e+00
#> 8          1       8 -1.628357761 -2.09184983 4.375836e+00
#> 9          1       9  1.099096129          NA           NA
#> 10         1      10  0.095564740          NA           NA
#> 11         2       1 -1.192322471 -1.20233784 1.445616e+00
#> 12         2       2 -0.292184020  0.63394460 4.018858e-01
#> 13         2       3  0.904031481  3.07422422 9.450855e+00
#> 14         2       4  0.378723623  2.00259619 4.010392e+00
#> 15         2       5  0.604929568  2.46405632 6.071574e+00
#> 16         2       6 -1.000137395 -0.81028029 6.565541e-01
#> 17         2       7  0.807282402  2.87685610 8.276301e+00
#> 18         2       8 -0.376117252  0.46272081 2.141105e-01
#> 19         2       9 -0.696751935          NA           NA
#> 20         2      10 -0.673842044          NA           NA
#> 21         3       1  0.798776048  2.85950314 8.176758e+00
#> 22         3       2  0.597404118  2.44870440 5.996153e+00
#> 23         3       3 -0.311625673  0.59428363 3.531730e-01
#> 24         3       4 -0.393228151  0.42781457 1.830253e-01
#> 25         3       5  2.188897121  5.69535013 3.243701e+01
#> 26         3       6  0.461360015  2.17117443 4.713998e+00
#> 27         3       7  0.974930068  3.21885734 1.036104e+01
#> 28         3       8  0.485507540  2.22043538 4.930333e+00
#> 29         3       9  1.448336223          NA           NA
#> 30         3      10  1.119655519          NA           NA
#> 31         4       1 -0.805235888 -0.41268121 1.703058e-01
#> 32         4       2  1.600780237  4.49559168 2.021034e+01
#> 33         4       3 -0.715547221 -0.22971633 5.276959e-02
#> 34         4       4 -1.287071150 -1.39562515 1.947770e+00
#> 35         4       5  0.815259874  2.89313014 8.370202e+00
#> 36         4       6 -0.070630992  1.08591278 1.179207e+00
#> 37         4       7 -0.018097109  1.19308190 1.423444e+00
#> 38         4       8 -0.970277248 -0.74936559 5.615488e-01
#> 39         4       9 -0.046794614          NA           NA
#> 40         4      10  0.876179522          NA           NA
#> 41         5       1 -0.045629610  1.13691559 1.292577e+00
#> 42         5       2  0.591829528  2.43733224 5.940588e+00
#> 43         5       3  0.390773389  2.02717771 4.109449e+00
#> 44         5       4  0.335713472  1.91485548 3.666672e+00
#> 45         5       5  1.356254073  3.99675831 1.597408e+01
#> 46         5       6 -0.214746801  0.79191653 6.271318e-01
#> 47         5       7 -0.012695749  1.20410067 1.449858e+00
#> 48         5       8 -0.928024047 -0.66316906 4.397932e-01
#> 49         5       9  1.953132106          NA           NA
#> 50         5      10  1.492505140          NA           NA
#> 51         6       1 -0.595075878  0.01604521 2.574487e-04
#> 52         6       2 -1.633478913 -2.10229698 4.419653e+00
#> 53         6       3  1.332074224  3.94743142 1.558221e+01
#> 54         6       4  0.186611837  1.61068815 2.594316e+00
#> 55         6       5 -0.008543471  1.21257132 1.470329e+00
#> 56         6       6  0.497079901  2.24404300 5.035729e+00
#> 57         6       7  0.083847694  1.40104930 1.962939e+00
#> 58         6       8  0.326840468  1.89675455 3.597678e+00
#> 59         6       9  0.098897888          NA           NA
#> 60         6      10  0.348056776          NA           NA
#> 61         7       1 -0.103945433  1.01795132 1.036225e+00
#> 62         7       2  0.227346734  1.69378734 2.868916e+00
#> 63         7       3  2.031012081  5.37326465 2.887197e+01
#> 64         7       4  0.221963908  1.68280637 2.831837e+00
#> 65         7       5  2.637643737  6.61079322 4.370259e+01
#> 66         7       6 -1.443755569 -1.71526136 2.942122e+00
#> 67         7       7 -0.956117286 -0.72047926 5.190904e-01
#> 68         7       8 -0.182490946  0.85771847 7.356810e-01
#> 69         7       9  1.861330816          NA           NA
#> 70         7      10  0.828443485          NA           NA
#> 71         8       1  0.531210254  2.31366892 5.353064e+00
#> 72         8       2 -0.004128931  1.22157698 1.492250e+00
#> 73         8       3 -0.676207567 -0.14946344 2.233932e-02
#> 74         8       4  0.228571453  1.69628576 2.877385e+00
#> 75         8       5  2.060793529  5.43401880 2.952856e+01
#> 76         8       6  1.250625379  3.78127577 1.429805e+01
#> 77         8       7  0.394004558  2.03376930 4.136218e+00
#> 78         8       8  0.542711034  2.33713051 5.462179e+00
#> 79         8       9 -0.305519502          NA           NA
#> 80         8      10  2.508846475          NA           NA
#> 81         9       1  0.713624674  2.68579434 7.213491e+00
#> 82         9       2  1.176272934  3.62959678 1.317397e+01
#> 83         9       3  1.220099799  3.71900359 1.383099e+01
#> 84         9       4 -0.701922247 -0.20192138 4.077225e-02
#> 85         9       5  0.498580322  2.24710386 5.049476e+00
#> 86         9       6  0.334212564  1.91179363 3.654955e+00
#> 87         9       7  1.082418318  3.43813337 1.182076e+01
#> 88         9       8 -1.268606521 -1.35795730 1.844048e+00
#> 89         9       9 -0.604883499          NA           NA
#> 90         9      10  0.242484199          NA           NA
#> 91        10       1  1.766816262  4.83430517 2.337051e+01
#> 92        10       2 -2.061966415 -2.97641149 8.859025e+00
#> 93        10       3 -0.991669740 -0.79300627 6.288589e-01
#> 94        10       4 -0.978416117 -0.76596888 5.867083e-01
#> 95        10       5 -0.209345980  0.80293420 6.447033e-01
#> 96        10       6  1.444200333  4.17616868 1.744038e+01
#> 97        10       7  0.144674259  1.52513549 2.326038e+00
#> 98        10       8 -2.126999077 -3.10907812 9.666367e+00
#> 99        10       9  0.464926452          NA           NA
#> 100       10      10  0.023903480          NA           NA

# Remove all rows above split_id
prepare_training_data(spacetime_data, y, split_id = 9, prep_strategy = "truncate")
#>    id_space id_time            x           y           y2
#> 1         1       1  0.846869397  2.95761357 8.747478e+00
#> 2         1       2  0.289797103  1.82118609 3.316719e+00
#> 3         1       3 -0.455661138  0.30045128 9.027097e-02
#> 4         1       4 -0.120609164  0.98395730 9.681720e-01
#> 5         1       5  0.323701479  1.89035102 3.573427e+00
#> 6         1       6 -0.242792754  0.73470278 5.397882e-01
#> 7         1       7 -1.162309122 -1.14111061 1.302133e+00
#> 8         1       8 -1.628357761 -2.09184983 4.375836e+00
#> 11        2       1 -1.192322471 -1.20233784 1.445616e+00
#> 12        2       2 -0.292184020  0.63394460 4.018858e-01
#> 13        2       3  0.904031481  3.07422422 9.450855e+00
#> 14        2       4  0.378723623  2.00259619 4.010392e+00
#> 15        2       5  0.604929568  2.46405632 6.071574e+00
#> 16        2       6 -1.000137395 -0.81028029 6.565541e-01
#> 17        2       7  0.807282402  2.87685610 8.276301e+00
#> 18        2       8 -0.376117252  0.46272081 2.141105e-01
#> 21        3       1  0.798776048  2.85950314 8.176758e+00
#> 22        3       2  0.597404118  2.44870440 5.996153e+00
#> 23        3       3 -0.311625673  0.59428363 3.531730e-01
#> 24        3       4 -0.393228151  0.42781457 1.830253e-01
#> 25        3       5  2.188897121  5.69535013 3.243701e+01
#> 26        3       6  0.461360015  2.17117443 4.713998e+00
#> 27        3       7  0.974930068  3.21885734 1.036104e+01
#> 28        3       8  0.485507540  2.22043538 4.930333e+00
#> 31        4       1 -0.805235888 -0.41268121 1.703058e-01
#> 32        4       2  1.600780237  4.49559168 2.021034e+01
#> 33        4       3 -0.715547221 -0.22971633 5.276959e-02
#> 34        4       4 -1.287071150 -1.39562515 1.947770e+00
#> 35        4       5  0.815259874  2.89313014 8.370202e+00
#> 36        4       6 -0.070630992  1.08591278 1.179207e+00
#> 37        4       7 -0.018097109  1.19308190 1.423444e+00
#> 38        4       8 -0.970277248 -0.74936559 5.615488e-01
#> 41        5       1 -0.045629610  1.13691559 1.292577e+00
#> 42        5       2  0.591829528  2.43733224 5.940588e+00
#> 43        5       3  0.390773389  2.02717771 4.109449e+00
#> 44        5       4  0.335713472  1.91485548 3.666672e+00
#> 45        5       5  1.356254073  3.99675831 1.597408e+01
#> 46        5       6 -0.214746801  0.79191653 6.271318e-01
#> 47        5       7 -0.012695749  1.20410067 1.449858e+00
#> 48        5       8 -0.928024047 -0.66316906 4.397932e-01
#> 51        6       1 -0.595075878  0.01604521 2.574487e-04
#> 52        6       2 -1.633478913 -2.10229698 4.419653e+00
#> 53        6       3  1.332074224  3.94743142 1.558221e+01
#> 54        6       4  0.186611837  1.61068815 2.594316e+00
#> 55        6       5 -0.008543471  1.21257132 1.470329e+00
#> 56        6       6  0.497079901  2.24404300 5.035729e+00
#> 57        6       7  0.083847694  1.40104930 1.962939e+00
#> 58        6       8  0.326840468  1.89675455 3.597678e+00
#> 61        7       1 -0.103945433  1.01795132 1.036225e+00
#> 62        7       2  0.227346734  1.69378734 2.868916e+00
#> 63        7       3  2.031012081  5.37326465 2.887197e+01
#> 64        7       4  0.221963908  1.68280637 2.831837e+00
#> 65        7       5  2.637643737  6.61079322 4.370259e+01
#> 66        7       6 -1.443755569 -1.71526136 2.942122e+00
#> 67        7       7 -0.956117286 -0.72047926 5.190904e-01
#> 68        7       8 -0.182490946  0.85771847 7.356810e-01
#> 71        8       1  0.531210254  2.31366892 5.353064e+00
#> 72        8       2 -0.004128931  1.22157698 1.492250e+00
#> 73        8       3 -0.676207567 -0.14946344 2.233932e-02
#> 74        8       4  0.228571453  1.69628576 2.877385e+00
#> 75        8       5  2.060793529  5.43401880 2.952856e+01
#> 76        8       6  1.250625379  3.78127577 1.429805e+01
#> 77        8       7  0.394004558  2.03376930 4.136218e+00
#> 78        8       8  0.542711034  2.33713051 5.462179e+00
#> 81        9       1  0.713624674  2.68579434 7.213491e+00
#> 82        9       2  1.176272934  3.62959678 1.317397e+01
#> 83        9       3  1.220099799  3.71900359 1.383099e+01
#> 84        9       4 -0.701922247 -0.20192138 4.077225e-02
#> 85        9       5  0.498580322  2.24710386 5.049476e+00
#> 86        9       6  0.334212564  1.91179363 3.654955e+00
#> 87        9       7  1.082418318  3.43813337 1.182076e+01
#> 88        9       8 -1.268606521 -1.35795730 1.844048e+00
#> 91       10       1  1.766816262  4.83430517 2.337051e+01
#> 92       10       2 -2.061966415 -2.97641149 8.859025e+00
#> 93       10       3 -0.991669740 -0.79300627 6.288589e-01
#> 94       10       4 -0.978416117 -0.76596888 5.867083e-01
#> 95       10       5 -0.209345980  0.80293420 6.447033e-01
#> 96       10       6  1.444200333  4.17616868 1.744038e+01
#> 97       10       7  0.144674259  1.52513549 2.326038e+00
#> 98       10       8 -2.126999077 -3.10907812 9.666367e+00