Remove overlapping clusters from scan statistics

In addition to reporting the details of the most likely cluster (MLC), scanstatistics objects also contain detailed information about the likelihood of a cluster in all possible spatial zones and temporal durations. This information can be used to identify any independent secondary clusters that exist in the study area. When searching for these secondary clusters, we are generally interested in clusters that do not significantly overlap the primary cluster.

Usage

remove_overlapping_clusters(
  scan_result,
  zones,
  order_by,
  take_highest = TRUE,
  max_overlap_frac = 0
)

Arguments

scan_result

An object of class scanstatistic, such as generated by scan_bayes_negbin or scan_eb_poisson2.

zones

A list of integer vectors. Each vector corresponds to a single zone; its elements are the numbers of the locations in that zone.

order_by

The column to order the clusters by. Any column in scan_result$observed or scan_result$posteriors$window_posteriors is valid. Specified as bare text.

take_highest

Boolean. Should the clusters by ordered from highest to lowest (take_highest = TRUE) or from lowest to highest (take_highest = FALSE).

max_overlap_frac

Optional. What is the maximum overlap fraction that should be allowed between clusters? Defaults to 0.

Value

An object of class scanstatistic, with the observed or posteriors$window_posteriors

data frame modified to only contain non-overlapping clusters. Otherwise identical to the scan_result input.

Details

remove_overlapping_clusters goes through the information about the likelihood of a cluster in all possible spatial zones and temporal durations, and only keeps the information from non-overlapping spatial zones. Potential clusters are gone through from most-likely to least likely (as determined by order_by and take_highest). Information about a spatial zone and duration is retained if the fraction of spatial locations in the zone that are present in any already identified cluster is less than or equal to max_overlap_frac.

See also

Examples

library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scanres <- scan_eb_poisson2(NM_data, count, zones,
                            baseline_est, n_mcsim = 100)

# Order by cluster score
remove_overlapping_clusters(scanres, zones, order_by = score)
#> Data distribution:                Poisson
#> Type of scan statistic:           expectation-based
#> Setting:                          univariate
#> Number of locations considered:   32
#> Maximum duration considered:      19
#> Number of spatial zones:          144
#> Number of Monte Carlo replicates: 100
#> Monte Carlo P-value:              0.079
#> Gumbel P-value:                   NULL
#> Most likely event duration:       7
#> ID of locations in MLC:           1, 10, 26, 30, 32

# Order by mc-pvalue
remove_overlapping_clusters(scanres, zones, order_by = mc_pvalue,
                            take_highest = FALSE)
#> Data distribution:                Poisson
#> Type of scan statistic:           expectation-based
#> Setting:                          univariate
#> Number of locations considered:   32
#> Maximum duration considered:      19
#> Number of spatial zones:          144
#> Number of Monte Carlo replicates: 100
#> Monte Carlo P-value:              0.079
#> Gumbel P-value:                   NULL
#> Most likely event duration:       7
#> ID of locations in MLC:           1, 10, 26, 30, 32

# Order by mc-pvalue, allow some overlap
remove_overlapping_clusters(scanres, zones, order_by = mc_pvalue,
                            take_highest = FALSE,
                            max_overlap_frac = 0.2)
#> Data distribution:                Poisson
#> Type of scan statistic:           expectation-based
#> Setting:                          univariate
#> Number of locations considered:   32
#> Maximum duration considered:      19
#> Number of spatial zones:          144
#> Number of Monte Carlo replicates: 100
#> Monte Carlo P-value:              0.079
#> Gumbel P-value:                   NULL
#> Most likely event duration:       7
#> ID of locations in MLC:           1, 10, 26, 30, 32