Extract clusters identified by scanning alarm functions

Extract all space-time clusters, based on values calculated by a scan statistic. This can either be an absolute measure of intensity, such as the score calculate by scan_eb_poisson or a relative measure, such as the Monte-Carlo p-value.

Usage

report_top_clusters(
  scan_result,
  order_by,
  take_highest = TRUE,
  cutoff_val = NULL,
  max_reported = 10,
  min_reported = 1,
  single_duration = TRUE
)

Arguments

scan_result

An object of class scanstatistic, such as generated by scan_bayes_negbin or scan_eb_poisson2.

order_by

The column to order the clusters by. Any column in scan_result$observed or scan_result$posteriors$window_posteriors is valid. Specified as bare text.

take_highest

Boolean. Should the clusters by ordered from highest to lowest (take_highest = TRUE) or from lowest to highest (take_highest = FALSE).

cutoff_val

The cutoff value used for reporting clusters. If take_highest = TRUE, only clusters with values of order_by greater than cutoff_val will be reported; if take_highest = FALSE, only report clusters with values under cutoff_val. Defaults to NULL, where no cutoff value is applied.

max_reported

The maximum number of clusters to be reported. Defaults to 10.

min_reported

The minimum number of clusters to be reported. This number of clusters will always be reported, even if they do not meet the threshold set by cutoff_val.

single_duration

Boolean. If TRUE, remove clusters with the same spatial zone but different duration as another, more intense, cluster.

Value

A data frame containing information about the most likely clusters in the study area, with between min_reported and max_reported rows, and including columns zone, duration, and the column specified by order_by.

Details

scanstatistics objects contain detailed information about the likelihood of a cluster in all possible spatial zones and temporal durations. report_top_clusters uses this information to report the zones most likely to contain a cluster. The results can be customized using the cutoff_val, max_reported, min_reported, and single_duration options.

See also

Examples

library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scanres <- scan_eb_poisson2(NM_data, count, zones,
                            baseline_est, n_mcsim = 100)

report_top_clusters(scanres, score)
#> # A tibble: 10 x 6
#>     zone duration score relrisk action_level mc_pvalue
#>    <dbl>    <dbl> <dbl>   <dbl>        <dbl>     <dbl>
#>  1   136        7  7.26    1.27         7.26    0.0297
#>  2   116        7  6.82    1.28         6.82    0.0396
#>  3   135        7  6.66    1.26         6.66    0.0396
#>  4   126        7  6.01    1.27         6.01    0.0891
#>  5   127        7  5.88    1.26         5.88    0.0891
#>  6    68        6  5.56    1.65         5.56    0.109 
#>  7    71        9  5.45    1.21         5.45    0.139 
#>  8   142        7  5.43    1.26         5.43    0.139 
#>  9     4        7  5.40    1.23         5.40    0.139 
#> 10   106        7  5.36    1.24         5.36    0.149 

report_top_clusters(scanres, mc_pvalue,
                    take_highest = FALSE,
                    cutoff_val = 0.05)
#> # A tibble: 3 x 6
#>    zone duration score relrisk action_level mc_pvalue
#>   <dbl>    <dbl> <dbl>   <dbl>        <dbl>     <dbl>
#> 1   136        7  7.26    1.27         7.26    0.0297
#> 2   116        7  6.82    1.28         6.82    0.0396
#> 3   135        7  6.66    1.26         6.66    0.0396

# Always report at least 3 clusters:
report_top_clusters(scanres, mc_pvalue,
                    take_highest = FALSE,
                    cutoff_val = 0.05,
                    min_reported = 3)
#> # A tibble: 3 x 6
#>    zone duration score relrisk action_level mc_pvalue
#>   <dbl>    <dbl> <dbl>   <dbl>        <dbl>     <dbl>
#> 1   136        7  7.26    1.27         7.26    0.0297
#> 2   116        7  6.82    1.28         6.82    0.0396
#> 3   135        7  6.66    1.26         6.66    0.0396

# First remove overlapping clusters, then report:
remove_overlapping_clusters(scanres, zones,
                            order_by = mc_pvalue,
                            take_highest = FALSE,
                            max_overlap_frac = 0.2) %>%
  report_top_clusters(mc_pvalue, take_highest = FALSE,
                      cutoff_val = 0.05)
#> # A tibble: 1 x 6
#>    zone duration score relrisk action_level mc_pvalue
#>   <dbl>    <dbl> <dbl>   <dbl>        <dbl>     <dbl>
#> 1   136        7  7.26    1.27         7.26    0.0297