Calculate the CUSUM statistic over a set of spatial-temporal zones

Calculate the CUSUM alarm statistic over a large set of spatio-temporal zones, as described by Sonesson (2007). scan_cusum_poisson2 is a helpful wrapper around scan_cusum_poisson allowing data to be passed as a single long data frame, instead of as a series of wide matrices.

Usage

scan_cusum_poisson2(
  spacetime_data,
  outcome_col,
  zones,
  baseline_col,
  scaling = 1.5,
  n_mcsim = 10
)

Arguments

spacetime_data

A data frame or a data frame extension (e.g., a tibble) that includes columns `id_space` and `id_time` that are each a set of consecutive integers. The column `id_time` identifies the time point associated with each row, with lower numbers being earlier and higher numbers being later. The column `id_space` identifies the spatial location or area associated with each row, but no meaning or order is ascribed to the values.

outcome_col

The column in spacetime_data containing the outcome of interest, such as observed counts

zones

A list of integer vectors. Each vector corresponds to a single zone; its elements are the numbers of the locations in that zone.

baseline_col

The column in data containing baseline estimates of the outcome of interest. Optional if pop_col is provided, otherwise required.

scaling

A scalar number > 1. Represents the proportional increase in the number of cases beyond wide_baseline in the event of an outbreak. If wide_baseline is the number of expected cases under normal conditions, then scaling * wide_baseline is the number of expected cases in the event of an outbreak. Used to calculate the downward drift.

n_mcsim

A non-negative integer; the number of replicate scan statistics to generate in order to calculate a \(P\)-value.

Value

An object of class scanstatistics, containing information about the type of scan statistic, the most likely cluster (MLC), and calculated scan statistics for every zone and duration investigated.

See also

Examples

library(sf)
data("NM_data")
data("NM_county_sf")
zones <- create_zones(NM_county_sf, max_k = 5)
scan_cusum_poisson2(NM_data, outcome_col = count, zones,
                    baseline_col = baseline_est)
#> Data distribution:                Poisson
#> Type of scan statistic:           cusum-based
#> Setting:                          univariate
#> Number of locations considered:   32
#> Maximum duration considered:      19
#> Number of spatial zones:          144
#> Number of Monte Carlo replicates: 10
#> Monte Carlo P-value:              0.182
#> Gumbel P-value:                   NULL
#> Most likely event duration:       6
#> ID of locations in MLC:           15, 26

scan_cusum_poisson2(NM_data, outcome_col = count, zones,
                    baseline_col = baseline_est,
                    scaling = 1.25,
                    n_mcsim = 99)
#> Data distribution:                Poisson
#> Type of scan statistic:           cusum-based
#> Setting:                          univariate
#> Number of locations considered:   32
#> Maximum duration considered:      19
#> Number of spatial zones:          144
#> Number of Monte Carlo replicates: 99
#> Monte Carlo P-value:              0.01
#> Gumbel P-value:                   NULL
#> Most likely event duration:       7
#> ID of locations in MLC:           1, 10, 26, 30, 32