Change how spatial zones are represented

Scan-type alarm functions work by computing an alarm function over a very large number of spatial zones, where each zone consists of one of more neighboring spatial locations. Information about which spatial locations belong to which zone can be specified in the form of a list, where each entry in the list contains the indexes of the locations in a particular zone, or in the form of a matrix, where each row represents a zone and each column represents a location. A set of zones can be generated using the create_zones function, and conversion between the list- and matrix-representation of the zones is accomplished through the functions zones_to_key_matrix and key_matrix_to_zones.

Usage

zones_to_key_matrix(zones)

key_matrix_to_zones(key_matrix)

Arguments

zones

A list where each entry is a zone. Each entry is a vector containing the indices of the spatial locations in that zone.

key_matrix

A matrix, where each row is a zone and each column is a location. The entry [i, j] is 1 if location j is a part of zone i and 0 otherwise.

Value

For zones_to_key_matrix, a matrix. To increase speed, the matrix is represented as a sparse matrix from the Matrix package. For create_zones, and key_matrix_to_zones a list of zones.

See also

Examples

library("sf")
data("NM_county_sf")

zones <- create_zones(NM_county_sf, max_k = 5)
key_matrix <- zones_to_key_matrix(zones)

zones <- key_matrix_to_zones(key_matrix)