Check that a data frame is formatted as a valid spacetime data frame

The statsurv package represents spacetime data as ordinary data frames that follow some specific conventions. These data frames must contain the columns id_space and id_time, that are each a set of consecutive integers. The column id_time identifies the time point associated with each row of the data frame, with lower numbers being earlier and higher numbers being later. The column id_space identifies the spatial location or area associated with each row. The specific values of id_space are not assumed to have any meaning or order beyond identification.

Usage

validate_spacetime_data(df)

Arguments

df

The data frame to validate

Value

Throws an error if the data frame is not a valid spacetime data frame. Otherwise, returns df, invisibly.

Details

To be a valid spacetime data frame, all of the following must be true:

The data frame must contain the columns id_time and id_space.

The entires in each of those columns must be consecutive integers

All combinations of id_time and id_space must be associated with at most 1 row.

Examples

valid_data <- data.frame(id_space = c(1, 2, 1, 2),
id_time = c(1, 1, 2, 2),
y = c(0.4, 0.5, 0.7, 0.8))

validate_spacetime_data(valid_data)

if (FALSE) {
  invalid_data <- data.frame(id_space = c(1, 2.5, 1, 2.5),
                             id_time = c(1, 1, 2, 2),
                             y = c(-1, -2, -0.5, -8))
  #Gives an error
  validate_spacetime_data(invalid_data)
}