Divide data into overlapping windows

This function divides a dataframe into overlapping time windows, in preparation for repeatedly applying statistical surveillance methods. Each window consists of 0 or more data points to fit a statistical model on, and 0 or more points to generate predictions for.

Usage

window_idtime(
  spacetime_data,
  min_train,
  max_train = Inf,
  n_predict = 1,
  step = 1,
  split_spatial_locations = FALSE
)

Arguments

spacetime_data

A data frame or a data frame extension (e.g., a tibble) that includes columns `id_space` and `id_time` that are each a set of consecutive integers. The column `id_time` identifies the time point associated with each row, with lower numbers being earlier and higher numbers being later. The column `id_space` identifies the spatial location or area associated with each row, but no meaning or order is ascribed to the values.

min_train, max_train

The minimum and maximum number of time points (as identified by id_time) to include in the training data set. Times with fewer than this number of time points will not be included as windows. Later times will include more time points in the training data set, up to the maximum number identified by max_train.

n_predict

The number of time points (as identified by id_time) to have predictions generated for.

step

How often should a window be generated? Default is 1, so a window is generated at every value of id_time.

split_spatial_locations

Whether each spatial area (as identified by id_space) should be located in its own data window (if split_spatial_locations = TRUE) or if they should all be included in one window (if split_spatial_locations = "FALSE").

Value

A tibble with one row per data window, containing the columns id_time, id_space, split_id, and curr_data. id_time and id_space identify the time point defining the end of the time point and the spatial area or areas associated with the data window. split_id is the first time point in the data window in the prediction set. The data for each window is stored in the curr_data column. There is no difference in the windowed data between data in the training set and the prediction set, but that information is used by other statsurv

functions.

Examples

x = rnorm(100)
spacetime_data <- data.frame(id_space = rep(1:10, each = 10),
                             id_time = rep(1:10, 10),
                             x = x,
                             y = 2.04 * x + 1.23)

# Create overlapping groups of data with at least 4 time points
# and up to 8 time points in the
# training set, and 2 time points in the prediction set.
window_idtime(spacetime_data, min_train = 4, max_train = 8,
              n_predict = 2, split_spatial_locations = TRUE)
#> # A tibble: 50 x 4
#>    window_time_id window_space_id split_id curr_data       
#>             <int>           <int>    <dbl> <list>          
#>  1              6               1        5 <tibble [6 x 4]>
#>  2              6               2        5 <tibble [6 x 4]>
#>  3              6               3        5 <tibble [6 x 4]>
#>  4              6               4        5 <tibble [6 x 4]>
#>  5              6               5        5 <tibble [6 x 4]>
#>  6              6               6        5 <tibble [6 x 4]>
#>  7              6               7        5 <tibble [6 x 4]>
#>  8              6               8        5 <tibble [6 x 4]>
#>  9              6               9        5 <tibble [6 x 4]>
#> 10              6              10        5 <tibble [6 x 4]>
#> # ... with 40 more rows

# The same, but put each spatial area in its own window
window_idtime(spacetime_data, min_train = 4, max_train = 8,
              n_predict = 2, split_spatial_locations = TRUE)
#> # A tibble: 50 x 4
#>    window_time_id window_space_id split_id curr_data       
#>             <int>           <int>    <dbl> <list>          
#>  1              6               1        5 <tibble [6 x 4]>
#>  2              6               2        5 <tibble [6 x 4]>
#>  3              6               3        5 <tibble [6 x 4]>
#>  4              6               4        5 <tibble [6 x 4]>
#>  5              6               5        5 <tibble [6 x 4]>
#>  6              6               6        5 <tibble [6 x 4]>
#>  7              6               7        5 <tibble [6 x 4]>
#>  8              6               8        5 <tibble [6 x 4]>
#>  9              6               9        5 <tibble [6 x 4]>
#> 10              6              10        5 <tibble [6 x 4]>
#> # ... with 40 more rows